### 2023-10-20 #615 @osorensen
galamm: Generalized Additive Latent and Mixed Models
Submitting Author Name: Øystein Sørensen
Submitting Author Github Handle: @osorensen
Repository:  https://github.com/LCBC-UiO/galamm
Version submitted: 0.1.1.9000
Submission type: Stats
Badge grade: gold

Version accepted: TBD

---

```
Package: galamm
Title: Generalized Additive Latent and Mixed Models
Version: 0.1.1.9000
    person(given = "Øystein",
           family = "Sørensen",
           email = "oystein.sorensen@psykologi.uio.no",
    person(given = "Douglas", family = "Bates", role = "ctb"),
    person(given = "Ben", family = "Bolker", role = "ctb"),
    person(given = "Martin", family = "Maechler", role = "ctb"),
    person(given = "Allan", family = "Leal", role = "ctb"),
    person(given = "Fabian", family = "Scheipl", role = "ctb"),
    person(given = "Steven", family = "Walker", role = "ctb"),
    person(given = "Simon", family = "Wood", role = "ctb")
           )
Description: Estimates generalized additive latent and
    mixed models using maximum marginal likelihood,
    as defined in Sorensen et al. (2023)
    <doi:10.1007/s11336-023-09910-z>, which is an extension of Rabe-Hesketh and
    Skrondal (2004)'s unifying framework for multilevel latent variable
    modeling <doi:10.1007/BF02295939>. Efficient computation is done using sparse
    matrix methods, Laplace approximation, and automatic differentiation. The
    framework includes generalized multilevel models with heteroscedastic
    residuals, mixed response types, factor loadings, smoothing splines,
    crossed random effects, and combinations thereof. Syntax for model
    formulation is close to 'lme4' (Bates et al. (2015)
    <doi:10.18637/jss.v067.i01>) and 'PLmixed' (Rockwood and Jeon (2019)
    <doi:10.1080/00273171.2018.1516541>).
    lme4,
    Matrix,
    memoise,
    mgcv,
    nlme,
    Rcpp,
    Rdpack,
    stats
    Rcpp,
    RcppEigen
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    covr,
    gamm4,
    PLmixed,
VignetteBuilder: knitr, rmarkdown
RdMacros: Rdpack
NeedsCompilation: yes
SystemRequirements: C++17
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in [issue 614](614)

The target audience is applied statisticians and quantitative scientists, particularly those working on the social sciences. The package is motivated by longitudinal studies in cognitive neuroscience, but it is applicable wherever a measurement model (of factor analysis type) needs to be combined with hierarchical modeling.

This is the first implementation of the algorithm developed in [Sørensen, Fjell, and Walhovd (2023)](https://doi.org/10.1007/s11336-023-09910-z).

gold

1. "Compliance with a good number of standards beyond those identified as minimally necessary.": I have attempted to comply with all standards for regression software outlined in the Online Book for Statistical Software. I have used **srr** to point out which parts of the code I think address each of the standards.
2. "Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories.": I have tried to comply with all the standards in 6.1.1 - 6.1.5 of the Standards Chapter.
3. "Have a demonstrated generality of usage beyond one single envisioned use case.": The software supports generality of usage, and the vignettes describe several such use cases.

Running `autotest` gives some errors, but they were waived in the pre-review issue.

The package is on CRAN. I am aware that rOpenSci recommends waiting with submitting to CRAN, but the package has some users already, and having pre-compiled binaries on CRAN makes it easier for them to install it, rather than having to set up a toolchain required for install from source. I hence opted to send it to CRAN.
