### 2023-11-11 90ef7a1
removed unnecessary default constructor
