### 2023-11-19 e3e2e80
Marked classes 'final' and added back destructor
