### 2023-12-21 #615 @osorensen
👋 @noamross, I just wanted to ask: what's the status of this submission? Is rOpenSci interested in reviewing it?
