### 2024-04-03 fc5ca84
fixed bug and added example dataset
