+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
@@ -11,7 +11,7 @@ articles:
   optimization: optimization.html
   scaling: scaling.html
   semiparametric: semiparametric.html
-last_built: 2023-10-20T07:55Z
+last_built: 2024-04-04T08:23Z
 urls:
   reference: https://lcbc-uio.github.io/galamm/reference
   article: https://lcbc-uio.github.io/galamm/articles
