+++ data-raw/lifespan.R
@@ -123,7 +123,8 @@ lifespan <- tibble(
     y = case_when(
       domain == "execfun" ~ -(y - mean(y)) / sd(y),
       TRUE ~ y
-    )
+    ),
+    id = factor(id)
   ) %>%
   ungroup() %>%
   select(-nu, -timepoint) %>%
