+++ vignettes/glmm_factor.Rmd
@@ -347,11 +347,6 @@ In this case we can confirm that the `galamm` function is correctly implemented 
 ```r
 library(lme4)
 #> Loading required package: Matrix
-#> 
-#> Attaching package: 'lme4'
-#> The following object is masked from 'package:galamm':
-#> 
-#>     llikAIC
 count_mod_lme4 <- glmer(
   formula = y ~ lbas * treat + lage + v4 + (1 | subj),
   data = epilep,
@@ -395,7 +390,7 @@ summary(count_mod_lme4)
 #> (Intercept)   1.7937     0.1045  17.165  < 2e-16 ***
 #> lbas          0.8844     0.1308   6.764 1.34e-11 ***
 #> treat        -0.3349     0.1474  -2.272  0.02311 *  
-#> lage          0.4845     0.3458   1.401  0.16121    
+#> lage          0.4845     0.3458   1.401  0.16122    
 #> v4           -0.1611     0.0543  -2.966  0.00301 ** 
 #> lbas:treat    0.3382     0.2025   1.671  0.09479 .  
 #> ---
