### 2024-04-05 9aff4a8
fixed sign error in stroop data for lifespan
