+++ vignettes/glmm_factor.Rmd
@@ -347,6 +347,11 @@ In this case we can confirm that the `galamm` function is correctly implemented 
 ```r
 library(lme4)
 #> Loading required package: Matrix
+#> 
+#> Attaching package: 'lme4'
+#> The following object is masked from 'package:galamm':
+#> 
+#>     llikAIC
 count_mod_lme4 <- glmer(
   formula = y ~ lbas * treat + lage + v4 + (1 | subj),
   data = epilep,
