### 2024-04-25 #615 @mpadge
@osorensen We'd still like to work with you to get this through our review process. How about one of the following options:

1.  We put the submission on hold here, until you let us know when your paper has passed through review. We'll then re-start the official review process straight after then. Or,
2. We start the review process anyway. Our reviews are generally completed [within just a few months at most](https://ropensci-review-tools.github.io/dashboard/history.html#review-duration). Especially given our slow response thus far, we'd ensure that our process would be completed well before a typical manuscript review. (Some good examples for recent stats submissions are #550 and #571, both done in around 2 months.) All changes and software improvements during our review could be used to support your responses to manuscript reviews.

Note that if your submission is to _Journal of Statistical Software_, then our system has been developed in collaboration with their processes, and they would likely welcome you using the results of a review here to support their own process.
