### 2024-04-27 #615 @osorensen
Thanks @mpadge, I go for option 1 then, and will ping you here once I've got a final decision on the paper.
