+++ cran-comments.md
@@ -1,18 +1,13 @@
 ## Submission note
 
-This is an update to the package which contains bug fixes, improvements of 
-the code base, and more documentation and examples.
+This is a small update which adds a recently accepted software paper to the 
+recommended citation of the package.
 
 ## Test environments
 
-* R development version with sanitizer support via rocker/r-devel-san.
-* R 4.3.1 build with valgrind on a local Ubuntu 22.04.
-* R Mac Builder (https://mac.r-project.org/macbuilder/submit.html)
-* Windows, r-release, r-oldrelease, r-devel.
-* Local Apple M1 Max, on R4.3.3.
+* Local Apple M1 Mac, on R4.4.0
 * R-CMD-check via GitHub Actions on windows-latest, macOS-latest, 
   ubuntu-20.04 (release), and ubuntu-20.04 (devel).
-
 
 ## R CMD check results
 
