### 2024-08-26 #615 @osorensen
@maelle, the above seems to be a false positive. The standards are documented here:

https://github.com/LCBC-UiO/galamm/blob/578b52c8a7db11401d98f12a2c21b7fdccc3e37f/tests/testthat.R#L1-L25
