### 2024-08-26 #615 @adamhsparks
@osorensen, unfortunately, that's not a false positive. They are not properly documented so that the bot can find the standards.

The documentation for ROxygen2 needs to be in /R, not /tests for it to be valid, _e.g._, https://github.com/giovsaraceno/QuadratiK-package/blob/master/R/srr-stats-standards.R

If you just move it, and redocument it should be fine.
