### 2024-08-27 #615 @osorensen
@mpadge, I moved the documentation to `R/`, so hope it's fine now.
