### 2024-08-27 #615 @adamhsparks
> @osorensen Sorry, bot's not quite ready for your case. Yours is the first package we've had with standards in `tests/testthat.R`. That is of course perfectly okay, but we simply hadn't encountered it, and so didn't realise that it caused a failure on our side. I'll ping here when our system has been rebuilt with the fix.
>
> (And FYI @adamhsparks, "srrstats" tags can go pretty much anywhere at all, including README, tests, src, and R code.)

Sorry, my bad. I was just trying to help based on previous cases.
