### 2024-08-27 #615 @mpadge
@adamhsparks and @osorensen The system has now been rebuilt, and the bot comment above that "This package ... may be submitted" would no longer appear. The first "Error:" would first have to be addressed.
