### 2024-08-28 #615 @mpadge
@osorensen The standards haven't changed, but since that time we've added additional checks to identify packages which only document general standards, with no category-specific ones. The general package checks also now generate and link to the full report on stats standards compliance, like in [this example here](https://github.com/ropensci/software-review/issues/632#issuecomment-2089078124), so we'd explicitly see there also that compliance statements with regression standards were missing.

So before proceeding, we'd need you to comply with our [*Regression Standards*](https://stats-devguide.ropensci.org/standards.html#standards-regression). Previous experience suggests that'll require around one day's work or so. As always, please let us know if you've got any questions, and in particular, feel free to comment in any way you like regarding the standards, ideally in the stats-devguide repo, linked directly from the book. Thanks!
