+++ R/galamm.R
@@ -4,7 +4,7 @@
 #'   (GALAMMs), as described in
 #'   \insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm}.
 #'   The building blocks of these models are generalized additive mixed models
-#'   (GAMMs) \insertCite{woodGeneralizedAdditiveModels2017a}{galamm}, of which
+#'   (GAMMs) \insertCite{woodGeneralizedAdditiveModels2017}{galamm}, of which
 #'   generalized linear mixed models
 #'   \insertCite{breslowApproximateInferenceGeneralized1993,harvilleMaximumLikelihoodApproaches1977,hendersonBestLinearUnbiased1975,lairdRandomEffectsModelsLongitudinal1982}{galamm}
 #'   are special cases. GALAMMs extend upon GAMMs by allowing factor structures,
@@ -38,13 +38,14 @@
 #' @srrstats {G2.2} Assertions on the lengths of arguments are implemented in
 #'   galamm.
 #' @srrstats {G2.3,G2.3a} match.arg() used on "na.action" argument.
-#' @srrstats {G2.3,G2.3b} Arguments "family", "load.var", "factor", and the elements
-#'   of the "start" argument are case sensitive. This is stated in the
+#' @srrstats {G2.3,G2.3b} Arguments "family", "load.var", "factor", and the
+#'   elements of the "start" argument are case sensitive. This is stated in the
 #'   documentation below.
-#' @srrstats {G2.4,G2.4a} Internally, objects family_mapping, weights_mapping and
-#'   lambda_mapping_X are explicitly converted to integer using as.integer().
-#' @srrstats {G2.4,G2.4b} as.numeric() used multiple places throughout the code for
-#'   explicitly converting to continuous.
+#' @srrstats {G2.4,G2.4a} Internally, objects family_mapping, weights_mapping
+#'   and lambda_mapping_X are explicitly converted to integer using
+#'   as.integer().
+#' @srrstats {G2.4,G2.4b} as.numeric() used multiple places throughout the code
+#'   for explicitly converting to continuous.
 #' @srrstats {G2.6} If \code{lambda} is provided as a vector, it will be
 #'   converted to a matrix with a single column, and a message will be printed.
 #' @srrstats {G2.7} Both \code{tibble}s and \code{data.table}s are accepted in
@@ -65,35 +66,67 @@
 #' @srrstats {G2.16} \code{NaN}, \code{Inf}, or \code{-Inf} in \code{data}
 #'   causes an error. The same happens with such values in \code{lambda}.
 #' @srrstats {G5.2} All errors and warnings are tested.
-#' @srrstats {G5.2a} Every message produced within R code by stop(),
-#'   warning(), or message(), is unique.
-#' @srrstats {G5.2b} All stop(), warning(), and message() calls are tested,
-#'   as can be seen in the CodeCov report on GitHub.
+#' @srrstats {G5.2a} Every message produced within R code by stop(), warning(),
+#'   or message(), is unique.
+#' @srrstats {G5.2b} All stop(), warning(), and message() calls are tested, as
+#'   can be seen in the CodeCov report on GitHub.
 #' @srrstats {G5.3} Tests have explicit expectations about return objects.
-#' @srrstats {G5.4a} These are new methods, but they have been used in the
-#'   paper Sørensen, Fjell, and Walhovd (2023), in which extensive simulation
-#'   studies confirmed the correctness of the implementation. Furthermore, the
-#'   simulated datasets, which are documented in "R/data.R" and exported, have
-#'   known ground truth and we confirm in the vignettes that the obtained
-#'   estimates are close to the true values.
[…]
