+++ R/formula.R
@@ -1,6 +1,7 @@
 #' @title Extract formula from fitted galamm object
 #'
 #' @srrstats {G1.4} Function documented with roxygen2.
+#' @srrstats {RE4.4} The specification of the model, generally as a formula (via formula())
 #' @param x Object of class \code{galamm} returned from \code{\link{galamm}}.
 #' @param ... Optional arguments passed on to other methods. Currently not used.
 #'
