+++ R/plot_smooth.R
@@ -10,7 +10,7 @@ plot_smooth <- function(object, ...) {
 #' @description
 #' Plots smooth terms of a fitted \code{galamm} object. This function is a thin
 #' wrapper around \code{mgcv::plot.gam}
-#' \insertCite{woodGeneralizedAdditiveModels2017a}{galamm}.
+#' \insertCite{woodGeneralizedAdditiveModels2017}{galamm}.
 #'
 #' @aliases plot_smooth plot_smooth.galamm
 #' @export plot_smooth
