### 2024-09-16 05ef997
added print.galamm function
