+++ R/predict.galamm.R
@@ -4,6 +4,8 @@
 #' @srrstats {G2.1a} Expected data types provided for all inputs.
 #' @srrstats {G2.3,G2.3a} match.arg() used on "type" argument.
 #' @srrstats {G2.3,G2.3b} Argument "type" is case sensitive, which is documented.
+#' @srrstats {RE4.16} Submitting new groups yields population level prediction,
+#'   which is the correct prediction for new groups in these models.
 #'
 #' @description Predictions are given at the population level, i.e., with random
 #'   effects set to zero. For fitted models including random effects, see
