+++ NAMESPACE
@@ -17,6 +17,7 @@ S3method(plot,galamm)
 S3method(plot_smooth,galamm)
 S3method(predict,galamm)
 S3method(print,VarCorr.galamm)
+S3method(print,galamm)
 S3method(print,summary.galamm)
 S3method(ranef,galamm)
 S3method(residuals,galamm)
