### 2024-09-16 5276610
fixed bug in test setup
