+++ inst/include/autodiff/forward/dual/dual.hpp
@@ -7,7 +7,7 @@
 //
 // Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 //
-// Copyright (c) 2018-2022 Allan Leal
+// Copyright © 2018–2024 Allan Leal
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -479,7 +479,7 @@ template<typename Op, typename L, typename C, typename R>
 struct AuxDualOpType<TernaryExpr<Op, L, C, R>> { using type = Op; };
 
 template<typename L, typename R>
-constexpr auto auxCommonDualType()
+AUTODIFF_DEVICE_FUNC constexpr auto auxCommonDualType()
 {
     if constexpr (isArithmetic<L> && isArithmetic<R>)
         return CommonType<L, R>();
@@ -512,17 +512,17 @@ struct Dual
 
     G grad = {};
 
-    Dual()
+    AUTODIFF_DEVICE_FUNC constexpr Dual()
     {}
 
     template<typename U, Requires<isExpr<U> || isArithmetic<U>> = true>
-    Dual(U&& other)
+    AUTODIFF_DEVICE_FUNC Dual(U&& other)
     {
         assign(*this, std::forward<U>(other));
     }
 
     template<typename U, Requires<isExpr<U> || isArithmetic<U>> = true>
-    Dual& operator=(U&& other)
+    AUTODIFF_DEVICE_FUNC Dual& operator=(U&& other)
     {
         Dual tmp;
         assign(tmp, std::forward<U>(other));
@@ -531,7 +531,7 @@ struct Dual
     }
 
     template<typename U, Requires<isExpr<U> || isArithmetic<U>> = true>
-    Dual& operator+=(U&& other)
+    AUTODIFF_DEVICE_FUNC Dual& operator+=(U&& other)
     {
         Dual tmp;
         assign(tmp, std::forward<U>(other));
@@ -540,7 +540,7 @@ struct Dual
     }
 
     template<typename U, Requires<isExpr<U> || isArithmetic<U>> = true>
-    Dual& operator-=(U&& other)
+    AUTODIFF_DEVICE_FUNC Dual& operator-=(U&& other)
     {
         Dual tmp;
         assign(tmp, std::forward<U>(other));
@@ -549,7 +549,7 @@ struct Dual
     }
 
     template<typename U, Requires<isExpr<U> || isArithmetic<U>> = true>
-    Dual& operator*=(U&& other)
+    AUTODIFF_DEVICE_FUNC Dual& operator*=(U&& other)
     {
         Dual tmp;
         assign(tmp, std::forward<U>(other));
@@ -558,7 +558,7 @@ struct Dual
     }
 
     template<typename U, Requires<isExpr<U> || isArithmetic<U>> = true>
-    Dual& operator/=(U&& other)
+    AUTODIFF_DEVICE_FUNC Dual& operator/=(U&& other)
     {
         Dual tmp;
         assign(tmp, std::forward<U>(other));
@@ -568,15 +568,15 @@ struct Dual
 
     /// Convert this Dual number into a value of type @p U.
 #if defined(AUTODIFF_ENABLE_IMPLICIT_CONVERSION_DUAL) || defined(AUTODIFF_ENABLE_IMPLICIT_CONVERSION)
-    operator T() const { return val; }
+    AUTODIFF_DEVICE_FUNC operator T() const { return val; }
 
     template<typename U>
-    operator U() const { return static_cast<U>(val); }
+    AUTODIFF_DEVICE_FUNC operator U() const { return static_cast<U>(val); }
 #else
-    explicit operator T() const { return val; }
[…]
