+++ inst/include/autodiff/common/meta.hpp
@@ -7,7 +7,7 @@
 //
 // Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 //
-// Copyright (c) 2018-2022 Allan Leal
+// Copyright © 2018–2024 Allan Leal
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -33,6 +33,15 @@
 #include <cstddef>
 #include <tuple>
 #include <type_traits>
+
+#ifndef AUTODIFF_DEVICE_FUNC
+#ifdef AUTODIFF_EIGEN_FOUND
+    #include <Eigen/src/Core/util/Macros.h>
+    #define AUTODIFF_DEVICE_FUNC EIGEN_DEVICE_FUNC
+#else
+    #define AUTODIFF_DEVICE_FUNC
+#endif
+#endif
 
 namespace autodiff {
 namespace detail {
@@ -68,13 +77,13 @@ template<typename Tuple>
 constexpr auto TupleSize = std::tuple_size_v<std::decay_t<Tuple>>;
 
 template<typename Tuple>
-constexpr auto TupleHead(Tuple&& tuple)
+AUTODIFF_DEVICE_FUNC constexpr auto TupleHead(Tuple&& tuple)
 {
     return std::get<0>(std::forward<Tuple>(tuple));
 }
 
 template<typename Tuple>
-constexpr auto TupleTail(Tuple&& tuple)
+AUTODIFF_DEVICE_FUNC constexpr auto TupleTail(Tuple&& tuple)
 {
     auto g = [&](auto&&, auto&&... args) constexpr {
         return std::forward_as_tuple(args...);
@@ -91,7 +100,7 @@ struct Index
 };
 
 template<size_t i, size_t ibegin, size_t iend, typename Function>
-constexpr auto AuxFor(Function&& f)
+AUTODIFF_DEVICE_FUNC constexpr auto AuxFor(Function&& f)
 {
     if constexpr (i < iend) {
         f(Index<i>{});
@@ -100,19 +109,19 @@ constexpr auto AuxFor(Function&& f)
 }
 
 template<size_t ibegin, size_t iend, typename Function>
-constexpr auto For(Function&& f)
+AUTODIFF_DEVICE_FUNC constexpr auto For(Function&& f)
 {
     AuxFor<ibegin, ibegin, iend>(std::forward<Function>(f));
 }
 
 template<size_t iend, typename Function>
-constexpr auto For(Function&& f)
+AUTODIFF_DEVICE_FUNC constexpr auto For(Function&& f)
 {
     For<0, iend>(std::forward<Function>(f));
 }
 
 template<size_t i, size_t ibegin, size_t iend, typename Function>
-constexpr auto AuxReverseFor(Function&& f)
+AUTODIFF_DEVICE_FUNC constexpr auto AuxReverseFor(Function&& f)
 {
     if constexpr (i < iend)
     {
@@ -122,19 +131,19 @@ constexpr auto AuxReverseFor(Function&& f)
 }
 
 template<size_t ibegin, size_t iend, typename Function>
-constexpr auto ReverseFor(Function&& f)
+AUTODIFF_DEVICE_FUNC constexpr auto ReverseFor(Function&& f)
 {
     AuxReverseFor<ibegin, ibegin, iend>(std::forward<Function>(f));
 }
 
 template<size_t iend, typename Function>
-constexpr auto ReverseFor(Function&& f)
+AUTODIFF_DEVICE_FUNC constexpr auto ReverseFor(Function&& f)
 {
     ReverseFor<0, iend>(std::forward<Function>(f));
 }
 
[…]

[…]
