+++ vignettes/semiparametric.Rmd
@@ -15,7 +15,7 @@ vignette: >
 
 
 
-```r
+``` r
 library(galamm)
 library(gamm4)
 ```
@@ -31,7 +31,7 @@ We start by showing how `galamm` can be used to estimated generalized additive m
 The `cognition` dataset contains simulated data with measurements of abilities in three cognitive domains.
 
 
-```r
+``` r
 head(cognition)
 #>   id domain          x timepoint item trials          y
 #> 1  1      1 0.06475113         1   11      1 0.16788973
@@ -45,14 +45,14 @@ head(cognition)
 For this first example, we focus only on the first item measured for the first domain.
 
 
-```r
+``` r
 dat <- subset(cognition, domain == 1 & item == "11")
 ```
 
 Each subject in this dataset has been measured eight times, and we can plot the measurements as follows:
 
 
-```r
+``` r
 plot(dat$x, dat$y, type = "n", xlab = "x", ylab = "y")
 for(i in unique(dat$id)) {
   dd <- dat[dat$id == i, ]
@@ -74,14 +74,14 @@ where $f(x_{ij})$ is a smooth function to be estimated, $\eta_{j} \sim N(0, \psi
 This model can be estimated using `gamm4` as follows:
 
 
-```r
+``` r
 mod_gamm4 <- gamm4(y ~ s(x), random = ~ (1 | id), data = dat, REML = FALSE)
 ```
 
 The package `gamm4` uses `lme4` to fit the underlying model, and the resulting model has two components. `mod_gamm4$mer` contains the mixed model representation, whereas in `mod_gamm4$gam` the fixed and random effects corresponding to spline coefficients have been converted into single smooth terms. We can look at the model summary for each:
 
 
-```r
+``` r
 summary(mod_gamm4$mer)
 #> Linear mixed model fit by maximum likelihood  ['lmerMod']
 #> 
@@ -134,7 +134,7 @@ summary(mod_gamm4$gam)
 We can also plot the estimated smooth term:
 
 
-```r
+``` r
 plot(mod_gamm4$gam)
 ```
 
@@ -143,7 +143,7 @@ plot(mod_gamm4$gam)
 In contrast, invoking the `plot` function on the mixed model part gives us a diagnostic plot.
 
 
-```r
+``` r
 plot(mod_gamm4$mer)
 ```
 
@@ -153,14 +153,14 @@ plot(mod_gamm4$mer)
 With `galamm` we use similar argument, but the `random` specification is now part of the model formula.
 
 
-```r
+``` r
 mod <- galamm(y ~ s(x) + (1 | id), data = dat)
 ```
 
 As opposed to `gamm4`, `galamm` gives a single summary. As can be seen, smooth terms are both reported as random effects, and in a separate line under the header "Approximate significance of smooth terms:". Reassuringly, the results from fitting the model with `gamm4` and with `galamm` are essentially equally, even though they use somewhat different computational algorithms.
 
 
-```r
+``` r
 summary(mod)
 #> GALAMM fit by maximum marginal likelihood.
 #> Formula: y ~ s(x) + (1 | id)
@@ -193,7 +193,7 @@ summary(mod)
 The `plot` function now gives us a diagnostic plot, which by inspection can be seen to be almost identical to the plot produced from the mixed model part of the `gamm4` model.
 
 
-```r
+``` r
 plot(mod)
 ```
 
@@ -202,7 +202,7 @@ plot(mod)
 In order to plot the smooth term, we use `plot_smooth`.
 
 
-```r
+``` r
 plot_smooth(mod)
 ```
 
@@ -211,7 +211,7 @@ plot_smooth(mod)
[…]
