+++ vignettes/latent_observed_interaction.Rmd
@@ -15,7 +15,7 @@ vignette: >
 
 
 
-```r
+``` r
 library(galamm)
 ```
 
@@ -26,7 +26,7 @@ This vignette describes how `galamm` can be used to model interactions between l
 For this example we use the simulated `latent_covariates` dataset, of which the first six rows are displayed below:
 
 
-```r
+``` r
 head(latent_covariates)
 #>   id         type         x            y response
 #> 1  1 measurement1 0.2655087 -0.530999307        0
@@ -79,7 +79,7 @@ The structural model is simply $\eta = \zeta \sim N(0, \psi)$, where $\psi$ is i
 It can be instructive to start by considering a model in which we fix $\\lambda_{4} = 0$. This type of model would be estimated with the following code:
 
 
-```r
+``` r
 lambda <- matrix(c(1, NA, NA), ncol = 1)
 
 mod0 <- galamm(
@@ -94,7 +94,7 @@ mod0 <- galamm(
 In the data generating simulations, the true values were $\lambda_{1}=1$, $\lambda_{2} = 1.3$ and $\lambda_{3} = -0.3$. The former two are very well recovered, but the latter is too positive, which is likely due to us omitting the interaction $\lambda_{4}$, whose true value was 0.2.
 
 
-```r
+``` r
 summary(mod0)
 #> GALAMM fit by maximum marginal likelihood.
 #> Formula: y ~ type + x:response + (0 + loading | id)
@@ -168,14 +168,14 @@ $$
 We specify the factor interactions with a list, one for each row of `lambda`:
 
 
-```r
+``` r
 factor_interactions <- list(~ 1, ~ 1, ~ x)
 ```
 
 This specifies that for the first two rows, there are no covariates, but for the third row, we want a linear regression with $x$ as covariate. Next, we specify the loading matrix **without** the interaction parameter, i.e., we reuse the `lambda` object that was specified for `mod0` above. This lets us fit the model as follows:
 
 
-```r
+``` r
 mod <- galamm(
   formula = y ~ type + x:response + (0 + loading | id),
   data = latent_covariates,
@@ -189,7 +189,7 @@ mod <- galamm(
 A model comparison shows overwhelming evidence in favor of this model, which is not surprising since this is how the data were simulated.
 
 
-```r
+``` r
 anova(mod, mod0)
 #> Data: latent_covariates
 #> Models:
@@ -205,7 +205,7 @@ anova(mod, mod0)
 The summary also shows that the bias in $\lambda_{3}$ has basically disappeared, as it is up to -0.318 from -0.195, with the true value being -0.3. The interaction is estimated at 0.233, which is also very close to the true value 0.2. It should of course be noted here that the noise level in this simulated dataset was set unrealistically low, to let us confirm that the implementation itself is correct.
 
 
-```r
+``` r
 summary(mod)
 #> GALAMM fit by maximum marginal likelihood.
 #> Formula: y ~ type + x:response + (0 + loading | id)
@@ -244,14 +244,14 @@ summary(mod)
 We can also try to add interactions between the $x^{2}$ and $\eta$. We first update the formula in `factor_interactions`:
 
 
-```r
+``` r
 factor_interactions <- list(~ 1, ~ 1, ~ x + I(x^2))
 ```
 
 Then we fit the model as before:
 
 
-```r
+``` r
 mod2 <- galamm(
   formula = y ~ type + x:response + (0 + loading | id),
   data = latent_covariates,
@@ -265,7 +265,7 @@ mod2 <- galamm(
 As can be seen, the coefficient for this squared interaction is not significantly different from zero.
 
 
-```r
+``` r
 summary(mod2)
 #> GALAMM fit by maximum marginal likelihood.
 #> Formula: y ~ type + x:response + (0 + loading | id)
[…]
