+++ man/plot_smooth.galamm.Rd
@@ -63,6 +63,7 @@ plot_smooth(mod, shade = TRUE, rug = FALSE)
 Other summary functions: 
 \code{\link{anova.galamm}()},
 \code{\link{plot.galamm}()},
+\code{\link{print.galamm}()},
 \code{\link{print.summary.galamm}()},
 \code{\link{summary.galamm}()}
 }
