+++ man/plot.galamm.Rd
@@ -33,6 +33,7 @@ generic function.
 Other summary functions: 
 \code{\link{anova.galamm}()},
 \code{\link{plot_smooth.galamm}()},
+\code{\link{print.galamm}()},
 \code{\link{print.summary.galamm}()},
 \code{\link{summary.galamm}()}
 }
