+++ man/print.galamm.Rd
@@ -31,4 +31,12 @@ print(mod)
 \seealso{
 \code{\link[=summary.galamm]{summary.galamm()}} for the summary function and \code{\link[=print]{print()}} for the
 generic.
+
+Other summary functions: 
+\code{\link{anova.galamm}()},
+\code{\link{plot.galamm}()},
+\code{\link{plot_smooth.galamm}()},
+\code{\link{print.summary.galamm}()},
+\code{\link{summary.galamm}()}
 }
+\concept{summary functions}
