### 2024-09-16 #615 @osorensen
@mpadge @adamhsparks, I'm not sure if I'm allowed to run `check srr` myself, but I've added tags which confirm that the package complies with the regression standards now. They are all in the main branch of my repository. Could you please check if this is sufficient?
