### 2024-09-16 #615 @mpadge
@osorensen Sorry, that was a wee :bug: on our side. Should work now, but note that result will currently indicate that [RE4.15](https://stats-devguide.ropensci.org/standards.html#prediction-extrapolation-and-forecasting) is missing, so maybe add that standard first before calling?
