### 2024-09-17 6f6a137
fixed latex error in vignette
