### 2024-09-17 #615 @mpadge
@osorensen @adamhsparks The previous calls seem to have issued during our weekly system rebuild. Just rotten timing, but all should work fine from here on. And nice work @osorensen - 100% compliance!!
