### 2024-11-06 #615 @nicholasjclark
---

### Compliance with Standards

- [X] This package complies with a sufficient number of standards for a _gold_ badge
- [X] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

- No `@srrstatsNA` tags were issued, and I agree that all standards apply if the authors wish to achieve a gold badge

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

- I find the adherence to standards to be very high throughout

For packages aiming for silver or gold badges:

- [X] This package extends beyond minimal compliance with standards in the following ways: (please describe)

- It is very nice to see the explicit recognition of authors / maintainers of crucial packages on which this package is built as contributors in the Description. This is a great way to champion open source software

- In line with the above, there is extensive referencing throughout the package that will be very helpful for users to better navigate and understand the modeling landscape

- Function documentation is extensive and has clearly been written to help novel users get started

- The code follows a strict `tidyverse` style guide throughout, which makes navigating easy and consistent. The same is also applied to internal functions, which are extremely easy to navigate and understand

- It is easy to find relevant examples that showcase all arguments from the primary functions, particularly `galamm()`

- The vignettes do an excellent job of introducing the problems that this package attempts to address, and of showcasing how users can work through the package

- Running `goodpractice::gp()` only returns some warnings about avoiding code lines that are longer than 80 characters. In my view these are not essential

---

### General Review

- The opening paragraphs of the README, and the information in `?galamm-package`, do a good job of explaining the problem being addressed. I think that this R package indeed solves a challenging problem, so very well done

- I had no issues with installation

- [X] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING

- There is adequate information in `CONTRIBUTING.md` for how to contribute, which is much appreciated

- [X] The documentation is sufficient to enable general use of the package beyond one specific use case

- There is a very comprehensive set of vignettes available already, and the examples in the various functions are also well documented. I was able to quickly fit some models to my own data, which I take as a good sign that the documentation can help users adapt these models for their own use cases

#### Algorithms

- The `.cpp` functions are all very well documented. I am no expert in these algorithms so cannot comment on their design and efficiency, but the level of documentation would make that very straightforward for others that may wish to help drive any possible improvements in future

- The tests of parameter recovery, and of the scaling of the precision of this recovery with data size, are valuable and informative

- All tests passed on my Windows machine, once I installed the necessary `PLmixed` package. In addition, all examples also passed. Code coverage is very high, which is impressive

- I confirmed that error messages from `mgcv` were passed on to the user (in my case I purposely set `k` too large). But explicit tests that key warnings or errors are passed on would potentially be valuable to help users design and build relevant models

- When erroring on the dimensions of lambda, I believe the message "lambda matrix must contain one row for each element in load.var" could be changed to "lambda matrix must contain one row for each *unique* element in load.var"

#### Visualisation (where appropriate)

- The visualisations are all appropriate and in line with standard practice for regression packages that use semiparametric functional predictors. However as I mentioned above, I do feel that consideration of making objects of class `galamm` available for use in `gratia`, `marginaleffects` or other popular post-processing packages would greatly extend the user's ability to interact with and interrogate their models

#### Package Design
- The design of the package is very familiar to users of popular regression modeling packages such as `lme4` or `mgcv`, so I feel that it won't be challenging for new users to understand how to navigate the functional landscape. It is exceptionally helpful to have common methods such as `anova()`, `ranef()`, `deviance()` and `predict()`. Other methods to perhaps consider in future would be `update()` or `tidy()`, depending on the directions the package may be going

- I find it confusing that the `weights` argument in `galamm()` is specified as a `formula`. This goes against convention for most regression modeling packages in R, where `weights` must be supplied as a `vector`. Perhaps the name of this argument should be changed to avoid confusion?

- Similarly I find the argument `load.var` to be inconsistent with the snake_case used throughout the rest of the `galamm()` function. Perhaps this should be changed to `load_var` instead?

- The `family_mapping` argument description does not specify that this needs to be a `vector` of class `integer`, so it is a bit confusing about how this should be set up. For example, I ran:
```r
dat <- head(subset(cognition, domain == 1 & timepoint == 1), 100)
loading_matrix <- matrix(c(1, NA, NA), ncol = 1)

mod <- galamm(
  formula = y ~ 0 + item + sl(x, k = 5, factor = "loading"),
  data = dat,
  load.var = "item",
  lambda = loading_matrix,
  family = c(gaussian, poisson),
  family_mapping = sample(c('a', 'b'), size = NROW(dat), replace = TRUE),
  factor = "loading"
)
```
And received the following feedback:

```r
Error in galamm(formula = y ~ 0 + item + sl(x, k = 5, factor = "loading"),  :
  family_mapping must contain a unique index for each element in family_list.
In galamm(formula = y ~ 0 + item + sl(x, k = 5, factor = "loading"),  :
  NAs introduced by coercion
```
It is not clear what `family_list` is here, and the NAs are introduced when trying to convert my string using `family_mapping <- as.integer(family_mapping)`. It may be useful to catch this type of behaviour earlier in the `galamm()` function

- Considering inter-operability, it may also be helpful to add methods (or request them via a PR) for the popular `gratia` and `marginaleffects` packages. It seems this would be easy enough given that an object of class `galamm` contains a `gam` object for the smooth outputs and that it already has a useful `predict` function. But I believe these additions would be helpful as both `gratia` and `marginaleffects` have support for generating `ggplot` outputs from many other objects that contain `mgcv`-style smooth outputs, as well as for testing particular hypotheses via model-based predictions

- I feel that more comprehensive sets of related functions could be listed in the `@seealso` tags, particularly for the primary function `galamm()`. As a user of new modeling packages I find it very helpful if these tags provide a general overview of some of the directions I might take after fitting a model. Obvious functions worth listing would be `summary.galamm()`, `plot.galamm()`, `factor_loadings()` and `anova.galamm()`.

- It may also be helpful to consider a help / MAN page that is dedicated to modeling strategies. This is an extremely flexible modeling package and the vignettes are great, but rather than working through specific examples you may want to give some more general overview of what the package can do without going into too many details. Have a look at `?mgcv::gam.models` as an example, and perhaps consider including guidance on how to maintain identifiability of the factors (i.e. are there recommendations about how many slots to fix and how many can be freely estimated?)

- I realize this may be too optimistic or possibly irrelevant, but the `mgcv` package can handle `list` datasets for setting up linear functional terms (i.e. [distributed lags](https://ecogambler.netlify.app/blog/distributed-lags-mgcv/), for example). I could envisage examples where users may want to model these types of nonlinear functions as well within the `galamm` framework, so it may be worth investigating whether they could also be supported

---

- The guidelines are strictly adhered to throughout the package
