### 2024-11-11 #615 @osorensen
Thanks for your review @nicholasjclark. @noamross, may I start addressing the points immediately, or is it better to wait until the second reviewer has completed his review also?
