### 2024-11-16 #615 @noamross
Thanks for your review, @nicholasjclark!  @osorensen, you can address things as you want, but I would suggest waiting for @eric-pedersen's review in case there's anything to be reconciled between the two.
