### 2025-02-02 #615 @maurolepore
@eric-pedersen, I just started my EiC rotation and I see your review was due on 23-November-2024. Are you still able to review?

No pressure. I totally understand things can change and you may be unable to continue. Just let us know so we decide how to move forward.
