### 2025-05-04 #615 @maurolepore
@osorensen I'm wrapping up my EiC roation so I'll summarize what I see to hand over this thread to the next EiC:

* This is a stats submission.
* One review has been submitted.
* The second review by @eric-pedersen was due on 21-March-2025.
* The author intended to incorporate feedback but is waiting for the second review.

@osorensen thanks a lot for sharing your work with rOpenSci.
