### 2025-05-10 e092b15
fixing uninitialized values
