### 2025-05-11 b5989a5
added step for fs to compile
