+++ .github/workflows/rhub.yaml
@@ -52,6 +52,12 @@ jobs:
 
     steps:
       - uses: r-hub/actions/checkout@v1
+
+      - name: Install system dependencies
+        run: |
+          apt-get update
+          apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev
+
       - uses: r-hub/actions/platform-info@v1
         with:
           token: ${{ secrets.RHUB_TOKEN }}
