### 2025-05-11 3426261
the moment you realize thinking would be better than relying on chatgpt
