### 2025-05-11 6611bba
adding r install
