+++ .github/workflows/rhub.yaml
@@ -51,10 +51,14 @@ jobs:
     steps:
       - uses: r-hub/actions/checkout@v1
 
+      - name: Install R
+        uses: r-lib/actions/setup-r@v2
+
       - name: Install system dependencies
         run: |
           sudo apt-get update
-          sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libgit2-dev libharfbuzz-dev libfribidi-dev
+          sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev
+
 
 
 
