### 2025-05-11 eb4c306
ready for CRAN
