+++ cran-comments.md
@@ -1,11 +1,11 @@
 ## Submission note
 
-This is a small update which adds a recently accepted software paper to the 
-recommended citation of the package.
+This is a fix of a WARNING on r-devel-linux-x86_64-fedora-gcc. In addition, some
+fixes to summary and print functions.
 
 ## Test environments
 
-* Local Apple M1 Mac, on R4.4.0
+* Local Apple M1 Mac, on R4.5.0
 * R-CMD-check via GitHub Actions on windows-latest, macOS-latest, 
   ubuntu-20.04 (release), and ubuntu-20.04 (devel).
 
