+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: galamm
 Title: Generalized Additive Latent and Mixed Models
-Version: 0.2.1.9000
+Version: 0.2.2
 Authors@R: c(
     person(given = "Øystein",
            family = "Sørensen",
