+++ cran-comments.md
@@ -8,6 +8,7 @@ I have (1) reproduced the original WARNING and (2) verified that this fix remove
 ## Test environments
 
 * Local Apple M1 Mac, on R4.5.0
+* Local gcc15 Ubuntu exactly reproducing r-devel-linux-x86_64-fedora-gcc on which there were warnings.
 * R-CMD-check via GitHub Actions on windows-latest, macOS-latest, 
   ubuntu-20.04 (release), and ubuntu-20.04 (devel).
 
