+++ vignettes/semiparametric.Rmd
@@ -85,8 +85,8 @@ The package `gamm4` uses `lme4` to fit the underlying model, and the resulting m
 summary(mod_gamm4$mer)
 #> Linear mixed model fit by maximum likelihood  ['lmerMod']
 #> 
-#>      AIC      BIC   logLik deviance df.resid 
-#>   3025.2   3052.1  -1507.6   3015.2     1595 
+#>       AIC       BIC    logLik -2*log(L)  df.resid 
+#>    3025.2    3052.1   -1507.6    3015.2      1595 
 #> 
 #> Scaled residuals: 
 #>      Min       1Q   Median       3Q      Max 
@@ -254,8 +254,8 @@ summary(mod_gamm4$mer)
 #> Generalized linear mixed model fit by maximum likelihood (Laplace Approximation) ['glmerMod']
 #>  Family: binomial  ( logit )
 #> 
-#>      AIC      BIC   logLik deviance df.resid 
-#>    983.7   1005.2   -487.8    975.7     1596 
+#>       AIC       BIC    logLik -2*log(L)  df.resid 
+#>     983.7    1005.2    -487.8     975.7      1596 
 #> 
 #> Scaled residuals: 
 #>     Min      1Q  Median      3Q     Max 
@@ -270,7 +270,7 @@ summary(mod_gamm4$mer)
 #> Fixed effects:
 #>              Estimate Std. Error z value Pr(>|z|)    
 #> X(Intercept)   2.8115     0.2034  13.825  < 2e-16 ***
-#> Xs(x)Fx1      -1.4110     0.4242  -3.327 0.000879 ***
+#> Xs(x)Fx1      -1.4110     0.4242  -3.326  0.00088 ***
 #> ---
 #> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
 #> 
@@ -603,20 +603,20 @@ mod_byvar <- galamm(
 #> At X0, 0 variables are exactly at the bounds
 #> At iterate     0  f=        25324  |proj g|=       9642.6
 #> At iterate    50  f =        136.9  |proj g|=        716.66
-#> At iterate   100  f =      -1054.7  |proj g|=        1485.1
-#> At iterate   150  f =      -1069.1  |proj g|=        29.652
-#> At iterate   200  f =      -1149.1  |proj g|=        920.26
+#> At iterate   100  f =      -1062.4  |proj g|=        356.24
+#> At iterate   150  f =      -1069.1  |proj g|=        19.086
+#> At iterate   200  f =      -1147.6  |proj g|=         10.73
 #> 
-#> iterations 237
-#> function evaluations 262
-#> segments explored during Cauchy searches 240
+#> iterations 201
+#> function evaluations 223
+#> segments explored during Cauchy searches 204
 #> BFGS updates skipped 0
 #> active bounds at final generalized Cauchy point 0
-#> norm of the final projected gradient 12.214
-#> final function value -1217.78
+#> norm of the final projected gradient 24.1617
+#> final function value -1147.63
 #> 
-#> F = -1217.78
-#> final  value -1217.784476 
+#> F = -1147.63
+#> final  value -1147.630941 
 #> converged
 ```
 
@@ -626,50 +626,49 @@ The summary shows that we have recovered the true values of the factor loadings 
 ``` r
 summary(mod_byvar)
 #> GALAMM fit by maximum marginal likelihood.
-#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability3")) +  
-#>     (0 + domain1:ability1 + domain3:ability3 | id/timepoint)
[…]
