### 2025-07-03 57b2d5a
Update R/galamm.R
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
