### 2025-07-03 1a426ce
Update tests/testthat/test-galamm-heteroscedastic.R
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
