### 2025-07-03 c92c42d
Update tests/testthat/test-galamm-heteroscedastic.R
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
