+++ cran-comments.md
@@ -7,6 +7,7 @@ This submission fixes a bug for heteroscedastic models with more than two groups
 * Local Apple M1 Mac, on R4.5.1
 * R-CMD-check via GitHub Actions on windows-latest, macOS-latest, 
   ubuntu-20.04 (release), and ubuntu-20.04 (devel).
+* win-builder (devel/release/oldrelease)
 
 ## R CMD check results
 
