### 2025-08-13 LCBC-UiO/galamm#200 @osorensen
Test that key warnings are passed on
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> I confirmed that error messages from mgcv were passed on to the user (in my case I purposely set k too large). But explicit tests that key warnings or errors are passed on would potentially be valuable to help users design and build relevant models
