### 2025-08-13 LCBC-UiO/galamm#201 @osorensen
Error message on lambda matrix
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> When erroring on the dimensions of lambda, I believe the message "lambda matrix must contain one row for each element in load.var" could be changed to "lambda matrix must contain one row for each unique element in load.var"
