### 2025-08-13 LCBC-UiO/galamm#202 @osorensen
Visualization
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> The visualisations are all appropriate and in line with standard practice for regression packages that use semiparametric functional predictors. However as I mentioned above, I do feel that consideration of making objects of class galamm available for use in gratia, marginaleffects or other popular post-processing packages would greatly extend the user's ability to interact with and interrogate their models
