### 2025-08-13 ropensci/galamm#203 @osorensen
rOpenSci review
Here is an overview of the issues from the rOpenSci review.

The following points were made by @nicholasjclark:
- [x] #200
- [x] #201
- [x] #202
- [x] #204
- [x] #205
- [x] #206
- [x] #207
- [ ] #208
- [x] #209

The following points were made by @dill:
- [x] #210
- [x] #211
- [x] #212
- [x] #213
- [x] #214
- [x] #215
- [x] #216
- [x] #217
- [x] #218
- [x] #219
- [x] #220
