### 2025-08-13 LCBC-UiO/galamm#204 @osorensen
Weights argument
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> I find it confusing that the weights argument in galamm() is specified as a formula. This goes against convention for most regression modeling packages in R, where weights must be supplied as a vector. Perhaps the name of this argument should be changed to avoid confusion?
