### 2025-08-13 LCBC-UiO/galamm#205 @osorensen
Snake case for consistency
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> Similarly I find the argument load.var to be inconsistent with the snake_case used throughout the rest of the galamm() function. Perhaps this should be changed to load_var instead?
