### 2025-08-13 LCBC-UiO/galamm#206 @osorensen
Family mapping function
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906):

The family_mapping argument description does not specify that this needs to be a vector of class integer, so it is a bit confusing about how this should be set up. For example, I ran:

dat <- head(subset(cognition, domain == 1 & timepoint == 1), 100)
loading_matrix <- matrix(c(1, NA, NA), ncol = 1)

mod <- galamm(
  formula = y ~ 0 + item + sl(x, k = 5, factor = "loading"),
  data = dat,
  load.var = "item",
  lambda = loading_matrix,
  family = c(gaussian, poisson),
  family_mapping = sample(c('a', 'b'), size = NROW(dat), replace = TRUE),
  factor = "loading"
)
And received the following feedback:

Error in galamm(formula = y ~ 0 + item + sl(x, k = 5, factor = "loading"),  :
  family_mapping must contain a unique index for each element in family_list.
In addition: Warning message:
In galamm(formula = y ~ 0 + item + sl(x, k = 5, factor = "loading"),  :
  NAs introduced by coercion
It is not clear what family_list is here, and the NAs are introduced when trying to convert my string using family_mapping <- as.integer(family_mapping). It may be useful to catch this type of behaviour earlier in the galamm() function
