### 2025-08-13 LCBC-UiO/galamm#207 @osorensen
Inter-operability
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> Considering inter-operability, it may also be helpful to add methods (or request them via a PR) for the popular gratia and marginaleffects packages. It seems this would be easy enough given that an object of class galamm contains a gam object for the smooth outputs and that it already has a useful predict function. But I believe these additions would be helpful as both gratia and marginaleffects have support for generating ggplot outputs from many other objects that contain mgcv-style smooth outputs, as well as for testing particular hypotheses via model-based predictions
