### 2025-08-13 LCBC-UiO/galamm#208 @osorensen
Modeling strategies help page
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> It may also be helpful to consider a help / MAN page that is dedicated to modeling strategies. This is an extremely flexible modeling package and the vignettes are great, but rather than working through specific examples you may want to give some more general overview of what the package can do without going into too many details. Have a look at ?mgcv::gam.models as an example, and perhaps consider including guidance on how to maintain identifiability of the factors (i.e. are there recommendations about how many slots to fix and how many can be freely estimated?)
