### 2025-08-13 LCBC-UiO/galamm#209 @osorensen
List data
@nicholasjclark wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-2458739906)

> I realize this may be too optimistic or possibly irrelevant, but the mgcv package can handle list datasets for setting up linear functional terms (i.e. [distributed lags](https://ecogambler.netlify.app/blog/distributed-lags-mgcv/), for example). I could envisage examples where users may want to model these types of nonlinear functions as well within the galamm framework, so it may be worth investigating whether they could also be supported
