### 2025-08-13 LCBC-UiO/galamm#210 @osorensen
S3 object documentation page
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> extensive documentation of return object in galamm is much appreciated! This is only a suggestion, but for mgcv Simon has a separate documentation object, gamObject, which just documents the return, you might want to move this stuff over there if there is a point where you feel like the galamm documentation is getting long (I don't feel like that's the case yet, just a note for future.)
