### 2025-08-13 LCBC-UiO/galamm#211 @osorensen
README
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> A statement of need clearly stating problems the software is designed to solve and its target audience in README

> Maybe link to the (very good) ["getting started"](https://lcbc-uio.github.io/galamm/articles/galamm.html) page?

> I might be tempted to include a (verbal) example at the very top of the README? I know that's difficult given the general nature of the software, but because of this the concepts are somewhat abstract. Given your users are likely to come from psychology (given the primary publication), perhaps a simple "for example, when we have data X..." type example would help solidify concepts? (Totally optional, just a suggestion.)
