### 2025-08-13 LCBC-UiO/galamm#212 @osorensen
Analytical derivatives
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> I couldn't exactly work out if this would work for you case (sorry my brain is not in C++ mode at the moment), but in mgcv I know that for [Wood, Pya and Säfken (2016)](https://www.tandfonline.com/doi/full/10.1080/01621459.2016.1180986) Simon included up to 4th derivatives of the likelihoods with analytical derivations in his code (and in the appendices for that paper). AD is fast, but these analytic derivatives may also be useful? I don't have enough experience with autodiff to know whether this is actually useful but I thought I'd mention it.
