### 2025-08-13 LCBC-UiO/galamm#213 @osorensen
Plot popping up in a test
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> galamm-semiparametric tests popped-up a plot, not clear if that should happen?
