### 2025-08-13 LCBC-UiO/galamm#214 @osorensen
plot function
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> plot.galamm seems to be the only function I found that was really lacking. The description doesn't tell the user what will be produced (just a plot, though the documentation uses the plural) or what to do with said plot (even a reference to standard literature would be useful to the naive user).
