### 2025-08-13 LCBC-UiO/galamm#215 @osorensen
Model checking
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> It would be worth talking about wider testing of assumptions somewhere, perhaps in individual vignettes?
