### 2025-08-13 LCBC-UiO/galamm#216 @osorensen
Pearson residuals, model checking
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> Perhaps a simplistic question, from someone who doesn't know the area: is there a reason that Pearson residuals are appropriate here? It would seem that probability inverse transform/randomized quantile residuals might be more appropriate in the Poisson/negbin/binomial response case? At least the deviance residuals would open up further response distribution checking (see next comment). Code from the statmod package might be useful here. See also [Smyth and Dunn (1996)](http://www.jstor.org/stable/1390802).
