### 2025-08-13 LCBC-UiO/galamm#217 @osorensen
plot.galamm and residuals
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> If you allow for other residuals, then the plot.galamm function needs to be able to specify the type of residuals.
