### 2025-08-13 LCBC-UiO/galamm#218 @osorensen
Model checking similar to mgcv
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> What about adding plots similar to those given by mgcv::gam.check ? These are relatively simple to construct for the generalized case and would give users a bit more information in terms of things like Q-Q plot and histogram of deviance residuals, plus plotting linear predictor against fitted values for a simple idea of the fit. A linear predictor vs. residuals can also be used to assess heteroskedasticy? Other plots such as looking at the variogram of residuals for spatial/temporal/longitudinal data might also be useful in this context. Simon Wood's GAM book has lots of examples of these types of checks (see for example, the brain imaging example and the Portuguese larks example).
