### 2025-08-13 LCBC-UiO/galamm#219 @osorensen
Better specification of multi-family models
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> I realise this may have not come to the author's attention previously, but regarding multi-likelihood models, this is possible in mgcv. This is via the gfam construct, which is well documented in the package. It may be worth adopting that standard for specifying such models, I find it a little cleaner that having the user write an ifelse (my R function nemesis).
