### 2025-08-13 LCBC-UiO/galamm#220 @osorensen
Posterior simulation
@dill wrote in his [rOpenSci review](https://github.com/ropensci/software-review/issues/615#issuecomment-3045619752)

> In your paper you talk about posterior simulation. I didn't see anything to that effect in the vignettes or documentation. Is it worth including some info on doing that? (Note that it may be worth cannibalizing mgcv::gam.mh for the generalized case.)
