### 2025-08-14 ab5fe1f
implemented two tests, closes #200
