+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: galamm
 Title: Generalized Additive Latent and Mixed Models
-Version: 0.2.3
+Version: 0.2.3.9000
 Authors@R: c(
     person(given = "Øystein",
            family = "Sørensen",
