+++ vignettes/mixed_response.Rmd
@@ -436,7 +436,7 @@ mod <- galamm(
 #> segments explored during Cauchy searches 58
 #> BFGS updates skipped 0
 #> active bounds at final generalized Cauchy point 0
-#> norm of the final projected gradient 0.0119052
+#> norm of the final projected gradient 0.0119457
 #> final function value 1372.16
 #> 
 #> F = 1372.16
