+++ vignettes/semiparametric.Rmd
@@ -269,7 +269,7 @@ summary(mod_gamm4$mer)
 #> 
 #> Fixed effects:
 #>              Estimate Std. Error z value Pr(>|z|)    
-#> X(Intercept)   2.8115     0.2034  13.825  < 2e-16 ***
+#> X(Intercept)   2.8115     0.2034  13.824  < 2e-16 ***
 #> Xs(x)Fx1      -1.4110     0.4242  -3.326  0.00088 ***
 #> ---
 #> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
@@ -436,7 +436,7 @@ summary(mod)
 #> 
 #> Scaled residuals: 
 #>    Min     1Q Median     3Q    Max 
-#>  0.693 10.402 26.734 35.728 51.878 
+#> -92.78 -65.29 -50.78 -17.21 -10.48 
 #> 
 #> Lambda:
 #>         loading       SE
@@ -457,7 +457,7 @@ summary(mod)
 #> item11            1.2694    0.06663 19.0513 6.412e-81
 #> item12            1.7788    0.09307 19.1128 1.977e-81
 #> item13            0.3797    0.02019 18.8077 6.531e-79
-#> s(x):loadingFx1  -0.1496    0.19977 -0.7488 4.540e-01
+#> s(x):loadingFx1   0.1496    0.19977  0.7488 4.540e-01
 #> 
 #> Approximate significance of smooth terms:
 #>                edf Ref.df    F p-value
@@ -603,20 +603,20 @@ mod_byvar <- galamm(
 #> At X0, 0 variables are exactly at the bounds
 #> At iterate     0  f=        25324  |proj g|=       9642.6
 #> At iterate    50  f =        136.9  |proj g|=        716.66
-#> At iterate   100  f =      -1062.4  |proj g|=        356.24
-#> At iterate   150  f =      -1069.1  |proj g|=        19.086
-#> At iterate   200  f =      -1147.6  |proj g|=         10.73
+#> At iterate   100  f =      -1062.4  |proj g|=        400.16
+#> At iterate   150  f =      -1069.1  |proj g|=        151.55
+#> At iterate   200  f =      -1147.5  |proj g|=        44.829
 #> 
-#> iterations 201
-#> function evaluations 223
-#> segments explored during Cauchy searches 204
+#> iterations 207
+#> function evaluations 228
+#> segments explored during Cauchy searches 210
 #> BFGS updates skipped 0
 #> active bounds at final generalized Cauchy point 0
-#> norm of the final projected gradient 24.1617
-#> final function value -1147.63
+#> norm of the final projected gradient 5.59584
+#> final function value -1147.55
 #> 
-#> F = -1147.63
-#> final  value -1147.630941 
+#> F = -1147.55
+#> final  value -1147.545485 
 #> converged
 ```
 
@@ -626,49 +626,50 @@ The summary shows that we have recovered the true values of the factor loadings 
 ``` r
 summary(mod_byvar)
 #> GALAMM fit by maximum marginal likelihood.
-#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability3")) +      (0 + domain1:ability1 + domain3:ability3 | id/timepoint)
+#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability3")) +  
+#>     (0 + domain1:ability1 + domain3:ability3 | id/timepoint)
 #>    Data: dat
 #> Control: galamm_control(optim_control = list(factr = 1e+09, trace = 3,      REPORT = 50, maxit = 1000))
 #> 
 #>      AIC      BIC   logLik deviance df.resid 
-#>  -2259.3  -2127.4   1147.6  -2295.3    11182 
[…]
