### 2025-08-22 LCBC-UiO/galamm#202 @osorensen
gratia functions actually work out of the box, as shown below:

``` r
library(galamm)
library(gamm4)
#> Loading required package: Matrix
#> Loading required package: lme4
#> Loading required package: mgcv
#> Loading required package: nlme
#>
#> Attaching package: 'nlme'
#> The following object is masked from 'package:lme4':
#>
#>     lmList
#> This is mgcv 1.9-3. For overview type 'help("mgcv-package")'.
#> This is gamm4 0.2-7
library(gratia)
dat <- subset(cognition, domain == 1)
loading_matrix <- matrix(c(1, NA, NA), ncol = 1)
mod <- galamm(
  formula = y ~ 0 + item + sl(x, factor = "loading") +
    (0 + loading | id / timepoint),
  data = dat,
  load_var = "item",
  lambda = loading_matrix,
  factor = "loading"
)

draw(mod$gam)
#> Warning: annotation$theme is not a valid theme.
#> Please use `theme()` to construct themes.
```

![](https://i.imgur.com/pAlsV2A.png)

``` r
derivatives(mod$gam)
#> # A tibble: 100 × 9
#>    .smooth      .by   .fs   .derivative   .se .crit .lower_ci .upper_ci        x
#>    <chr>        <chr> <chr>       <dbl> <dbl> <dbl>     <dbl>     <dbl>    <dbl>
#>  1 s(x):loading <NA>  <NA>         5.61 0.293  1.96      5.04      6.19 0.000829
#>  2 s(x):loading <NA>  <NA>         5.61 0.292  1.96      5.04      6.18 0.0109
#>  3 s(x):loading <NA>  <NA>         5.60 0.288  1.96      5.04      6.17 0.0210
#>  4 s(x):loading <NA>  <NA>         5.58 0.278  1.96      5.04      6.13 0.0311
#>  5 s(x):loading <NA>  <NA>         5.55 0.262  1.96      5.03      6.06 0.0412
#>  6 s(x):loading <NA>  <NA>         5.50 0.239  1.96      5.03      5.96 0.0513
#>  7 s(x):loading <NA>  <NA>         5.44 0.211  1.96      5.02      5.85 0.0614
#>  8 s(x):loading <NA>  <NA>         5.36 0.180  1.96      5.01      5.72 0.0715
#>  9 s(x):loading <NA>  <NA>         5.29 0.151  1.96      4.99      5.58 0.0816
#> 10 s(x):loading <NA>  <NA>         5.21 0.128  1.96      4.96      5.46 0.0917
#> # ℹ 90 more rows
appraise(mod$gam)
#> Warning: annotation$theme is not a valid theme.
#> Please use `theme()` to construct themes.
```

![](https://i.imgur.com/ZLhvXfl.png)

What I need to do, is to make thin wrappers of the type below:

```r
draw.galamm <- function(object, ...) {
  draw(object$gam, ...)
}
```

and similarly for derivatives and appraise.
