### 2025-09-26 329825e
added a bit more to readme
