### 2025-09-26 f35db2d
had to add a family to galammObject docs
