+++ man/galammObject.Rd
@@ -52,3 +52,10 @@ smooth terms.
 }
 }
 
+\seealso{
+Other modeling functions: 
+\code{\link{galamm}()},
+\code{\link{s}()},
+\code{\link{t2}()}
+}
+\concept{modeling functions}
