### 2025-09-26 LCBC-UiO/galamm#204 @osorensen
I agree. The reason for choosing the name "weights" in the first place is that this is the name of the similar argument in nlme. However, the fact that nlme has an unfortunate name is not a reason to continue the tradition. This has nothing to directly to do with weights really, since it's about setting up heteroscedastic residuals. Doing some research I found that glmmTMB uses the term `dispformula`, which is short for "dispersion formula". I think this is better, so will change it to that.
