### 2025-09-26 LCBC-UiO/galamm#204 @osorensen
In the PR soon to come, I have changed from "weights" to "dispformula", but I keep referring to the heteroscedasticity parameters as "weights". I think this makes sense, since the kind of models we have in mind here are of the type y = a + x * b + e, where var(e) = sigma if the observation belongs to group 1 and var(e) = W * sigma if the observation belongs to group 2, etc.
