### 2025-09-26 e45ea06
added deprecation warning and test for it
