+++ R/sigma.R
@@ -26,7 +26,7 @@
 #' # Linear mixed model with heteroscedastic residuals
 #' mod <- galamm(
 #'   formula = y ~ x + (1 | id),
-#'   weights = ~ (1 | item),
+#'   dispformula = ~ (1 | item),
 #'   data = hsced
 #' )
 #'
