+++ R/vcov.R
@@ -25,7 +25,7 @@
 #' # Linear mixed model with heteroscedastic residuals
 #' mod <- galamm(
 #'   formula = y ~ x + (1 | id),
-#'   weights = ~ (1 | item),
+#'   dispformula = ~ (1 | item),
 #'   data = hsced
 #' )
 #'
@@ -72,7 +72,7 @@ vcov.galamm <- function(object, parm = "beta", ...) {
 #' @examples
 #' mod <- galamm(
 #'   formula = y ~ x + (1 | id),
-#'   weights = ~ (1 | item),
+#'   dispformula = ~ (1 | item),
 #'   data = hsced
 #' )
 #'
