+++ vignettes/mixed_response.Rmd
@@ -154,7 +154,7 @@ This dataset has the same structure as `mresp`, except that the residual standar
 family_mapping <- ifelse(mresp$itemgroup == "a", 1L, 2L)
 mod <- galamm(
   formula = y ~ x + (1 | id),
-  weights = ~ (0 + isgauss | grp),
+  dispformula = ~ (0 + isgauss | grp),
   family = c(gaussian, binomial),
   family_mapping = family_mapping,
   data = mresp_hsced
