### 2025-09-26 LCBC-UiO/galamm#206 @osorensen
This is related to #219 and I should address them jointly.
