### 2025-09-26 c9a1f75
rebuilt long-running vignettes
