+++ vignettes/lmm_heteroscedastic.Rmd
@@ -64,7 +64,6 @@ summary(mod)
 #> GALAMM fit by maximum marginal likelihood.
 #> Formula: y ~ x + (1 | id)
 #>    Data: hsced
-#> Weights: ~(1 | item)
 #> 
 #>      AIC      BIC   logLik deviance df.resid 
 #>   4126.3   4151.7  -2058.1   4116.3     1195 
