### 2025-09-26 LCBC-UiO/galamm#219 @osorensen
@dill thanks for making me aware of the `gfam` argument from `mgcv`. I agree it is clearer, and will implementing something very similar in a PR to come soon.
