### 2025-09-26 4e9260e
made an implementation of gfam which works
