+++ R/galamm.R
@@ -152,28 +152,14 @@
 #' @param data A data.frame containing all the variables specified by the model
 #'   formula, with the exception of factor loadings.
 #'
-#' @param family A a list or character vector containing one or more model
-#'   families. For each element in \code{family} there should be a corresponding
-#'   element in \code{family_mapping} specifying which elements of the response
-#'   are conditionally distributed according to the given family. Currently
-#'   family can be one of \code{gaussian}, \code{binomial}, and \code{poisson},
-#'   and only canonical link functions are supported. The family arguments can
-#'   either be provided as character values, e.g., \code{c("gaussian",
-#'   "poisson")} or \code{list("gaussian", "poisson")}, as function names, e.g.,
-#'   \code{c(gaussian, poisson)} or \code{list(gaussian, poisson)}, or as
-#'   function calls, e.g., \code{list(gaussian(), poisson())}. In the latter
-#'   case, they must be provided in a list, and bot as a vector. Mixing the
-#'   different ways of describing the family also works, e.g.,
-#'   \code{list("gaussian", poisson())}, but in this case they must be provided
-#'   in a list. When provided as character values, the argument is case
-#'   sensitive.
+#' @param family A family object specifying the response distribution of the
+#'   model. Currently \code{gaussian}, \code{binomial}, and \code{poisson}
+#'   with canonical link functions are supported. Models with mixed response
+#'   types, in which responses are distributed according to one of the three
+#'   supported distributions, can be set up using the [gfam()] function.
 #'
-#'
-#' @param family_mapping Optional vector mapping from the elements of
-#'   \code{family} to rows of \code{data}. Defaults to \code{rep(1,
-#'   nrow(data))}, which means that all observations are distributed according
-#'   to the first element of \code{family}. The length of \code{family_mapping}
-#'   must be identical to the number of observations, \code{nrow(data)}.
+#' @param family_mapping Deprecated. See the documentation to [gfam()] for how
+#'   to set up mixed response type models.
 #'
 #' @param load_var Optional character specifying the name of the variable in
 #'   \code{data} identifying what the factors load onto. Default to \code{NULL},
@@ -320,7 +306,7 @@
 #'
 #' @md
 galamm <- function(formula, dispformula = NULL, weights = NULL, data,
-                   family = gaussian, family_mapping = rep(1, nrow(data)),
+                   family = gaussian, family_mapping = NULL,
                    load_var = NULL, load.var = NULL, lambda = NULL,
                    factor = NULL, factor_interactions = NULL,
                    na.action = getOption("na.action"),
@@ -331,8 +317,14 @@ galamm <- function(formula, dispformula = NULL, weights = NULL, data,
     if (is.null(load_var)) load_var <- load.var
   }
   if (!is.null(weights)) {
-    warning("`weights` is deprecated; use `dispformula` instead.", call. = FALSE)
+    warning("`weights` is deprecated; use `dispformula` instead.",
+            call. = FALSE)
     if (is.null(dispformula)) dispformula <- weights
+  }
+  if (!is.null(family_mapping)) {
+    warning("`family_mapping` is deprecated; see the gfam() function.",
+            call. = FALSE)
+    # Deal with the case here
   }
 
   if (!inherits(data, "data.frame")) {
[…]
