+++ R/misc.R
@@ -124,27 +124,29 @@ setup_family <- function(family) {
 #' @return A matrix with responses and number of trials.
 #' @noRd
 #'
-setup_response_object <- function(family_list, family_mapping, data, gobj) {
+setup_response_object <- function(family_list, data, gobj) {
   response_obj <- matrix(nrow = nrow(gobj$lmod$X), ncol = 2)
 
   for (i in seq_along(family_list)) {
     f <- family_list[[1]]
+    response_name <- all.vars(lme4::nobars(gobj$fake.formula))[[1]]
+    matching_rows <- data[, response_name][, 2] == i
     mf <- stats::model.frame(lme4::nobars(gobj$fake.formula),
-      data = data[family_mapping == i, ]
+      data = data[matching_rows, ]
     )
     mr <- stats::model.response(mf)
 
     if (f$family == "binomial" && !is.null(dim(mr))) {
       trials <- rowSums(mr)
     } else {
-      trials <- rep(1, sum(family_mapping == i))
+      trials <- rep(1, sum(matching_rows))
     }
     if (is.matrix(mr)) {
       response <- mr[, 1, drop = TRUE]
     } else {
       response <- mr
     }
-    response_obj[family_mapping == i, ] <-
+    response_obj[matching_rows, ] <-
       cbind(response = response, trials = trials)
   }
   response_obj
