+++ man/gfam.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gfam.R
+\name{gfam}
+\alias{gfam}
+\title{Grouped families}
+\usage{
+gfam(fl)
+}
+\arguments{
+\item{fl}{A list of families. Currently \code{gaussian}, \code{binomial}, and
+\code{poisson} with canonical link functions are supported.}
+}
+\value{
+An object of class "extended_family".
+}
+\description{
+This function is inspired by a function of the same name in the \code{mgcv}
+package \insertCite{woodGeneralizedAdditiveModels2017}{galamm}, and supports
+setting up mixed response type models. When using this function, the response
+variable must be supported as a two-column matrix, in which the first column
+contains the response and the second column contains an index mapping the
+response to the list elements provided in the argument \code{fl} to this function.
+}
+\references{
+\insertAllCited{}
+}
