+++ man/mresp_hsced.Rd
@@ -5,12 +5,14 @@
 \alias{mresp_hsced}
 \title{Simulated Mixed Response Data with Heteroscedastic Residuals}
 \format{
-\subsection{\code{mresp} A data frame with 4000 rows and 5 columns:}{
+\subsection{\code{mresp_hsced} A data frame with 4000 rows and 5 columns:}{
 
 \describe{
 \item{id}{Subject ID.}
 \item{x}{Predictor variable.}
-\item{y}{Response.}
+\item{y}{Matrix whose first column is the response and whose second
+column equals 1 if the observation is conditionally normally distributed
+and 2 if the observation is conditionally binomially distributed.}
 \item{itemgroup}{Factor variable which equals "a" for the normally
 distributed responses and "b" for the binomially distributed response
 (with 1 trial).}
