+++ NAMESPACE
@@ -30,6 +30,7 @@ export(factor_loadings)
 export(fixef)
 export(galamm)
 export(galamm_control)
+export(gfam)
 export(plot_smooth)
 export(ranef)
 export(response)

[…]
