+++ R/misc.R
@@ -100,7 +100,7 @@ setup_factor <- function(load_var, lambda, factor, data) {
 setup_family <- function(family) {
   if (length(family) == 1 || inherits(family, "family")) family <- list(family)
 
-  lapply(family, function(f) {
+  ret <- lapply(family, function(f) {
     if (is.character(f)) {
       return(eval(parse(text = f))())
     } else if (is.function(f)) {
@@ -109,6 +109,14 @@ setup_family <- function(family) {
       return(f)
     }
   })
+
+  lapply(ret, function(f) {
+    if(!inherits(f, "family")) {
+      stop("Not a family object")
+    }
+  })
+
+  ret
 }
 
 #' Create the response object
@@ -125,12 +133,25 @@ setup_family <- function(family) {
 #' @noRd
 #'
 setup_response_object <- function(family_list, data, gobj) {
-  response_obj <- matrix(nrow = nrow(gobj$lmod$X), ncol = 2)
+  response_obj <- matrix(nrow = nrow(gobj$lmod$X), ncol = 3)
+  response_name <- all.vars(lme4::nobars(gobj$fake.formula))[[1]]
+
+  if(length(family_list) > 1) {
+    if(!isTRUE(base::all.equal(
+      sort(as.integer(unique(data[, response_name][, 2]))),
+      seq_along(family_list)))) {
+      stop("There must be at least one index per family")
+    }
+  }
 
   for (i in seq_along(family_list)) {
     f <- family_list[[1]]
-    response_name <- all.vars(lme4::nobars(gobj$fake.formula))[[1]]
-    matching_rows <- data[, response_name][, 2] == i
+    matching_rows <- if(length(family_list) > 1) {
+      data[, response_name][, 2] == i
+    } else {
+      rep(TRUE, nrow(data))
+    }
+
     mf <- stats::model.frame(lme4::nobars(gobj$fake.formula),
       data = data[matching_rows, ]
     )
@@ -147,7 +168,7 @@ setup_response_object <- function(family_list, data, gobj) {
       response <- mr
     }
     response_obj[matching_rows, ] <-
-      cbind(response = response, trials = trials)
+      cbind(response = response, trials = trials, family = i)
   }
   response_obj
 }
