### 2025-09-27 cb16e39
updated vignettes run-extended
