+++ vignettes-raw/optimization.Rmd
@@ -44,7 +44,6 @@ We will illustrate some ways of modifying the optimization procedure with the co
 ```{r}
 loading_matrix <- matrix(c(1, 1, NA), ncol = 1)
 families <- c(gaussian, binomial)
-family_mapping <- ifelse(diet$item == "chd", 2, 1)
 formula <- y ~ 0 + chd + (age * bus):chd + fiber +
   (age * bus):fiber + fiber2 + (0 + loading | id)
 ```
@@ -56,7 +55,6 @@ mod <- galamm(
   formula = formula,
   data = diet,
   family = families,
-  family_mapping = family_mapping,
   factor = "loading",
   load_var = "item",
   lambda = loading_matrix
@@ -79,7 +77,6 @@ mod <- galamm(
   formula = formula,
   data = diet,
   family = families,
-  family_mapping = family_mapping,
   factor = "loading",
   load_var = "item",
   lambda = loading_matrix,
@@ -103,7 +100,6 @@ mod <- galamm(
   formula = formula,
   data = diet,
   family = families,
-  family_mapping = family_mapping,
   factor = "loading",
   load_var = "item",
   lambda = loading_matrix,
@@ -125,7 +121,6 @@ mod <- galamm(
   formula = formula,
   data = diet,
   family = families,
-  family_mapping = family_mapping,
   factor = "loading",
   load_var = "item",
   lambda = loading_matrix,
@@ -164,7 +159,6 @@ mod_nm <- galamm(
   formula = formula,
   data = diet,
   family = families,
-  family_mapping = family_mapping,
   factor = "loading",
   load_var = "item",
   lambda = loading_matrix,
