+++ tests/testthat/test-galamm-mixed-resp.R
@@ -175,7 +175,6 @@ test_that("Covariate measurement error model works", {
 
 
 test_that("Mixed response and heteroscedastic error works", {
-  family_mapping <- ifelse(mresp$itemgroup == "a", 1L, 2L)
   mod <- galamm(
     formula = y ~ x + (1 | id),
     dispformula = ~ (0 + isgauss | grp),
