### 2025-09-27 fc93de4
updating cognition data
