+++ data-raw/cognition.R
@@ -78,4 +78,6 @@ cognition <- crossing(
   ) %>%
   as.data.frame()
 
+cognition$y <- matrix(c(cognition$y, ifelse(cognition$domain == 2, 2, 1)), ncol = 2)
+
 usethis::use_data(cognition, overwrite = TRUE)
