+++ man/galamm.Rd
@@ -218,6 +218,7 @@ summary(mod)
 \seealso{
 Other modeling functions: 
 \code{\link{galammObject}},
+\code{\link{gfam}()},
 \code{\link{s}()},
 \code{\link{t2}()}
 }
