+++ man/gfam.Rd
@@ -24,3 +24,11 @@ response to the list elements provided in the argument \code{fl} to this functio
 \references{
 \insertAllCited{}
 }
+\seealso{
+Other modeling functions: 
+\code{\link{galamm}()},
+\code{\link{galammObject}},
+\code{\link{s}()},
+\code{\link{t2}()}
+}
+\concept{modeling functions}
