+++ man/galammObject.Rd
@@ -55,6 +55,7 @@ smooth terms.
 \seealso{
 Other modeling functions: 
 \code{\link{galamm}()},
+\code{\link{gfam}()},
 \code{\link{s}()},
 \code{\link{t2}()}
 }
