### 2025-09-28 2a2a57d
added a check and a test to avoid confusion with mgcv::gfam
