+++ R/galamm.R
@@ -348,7 +348,7 @@ galamm <- function(formula, dispformula = NULL, weights = NULL, data,
     stop("dispformula must be a formula")
   }
 
-  if(inherits(family, "extended.family")) {
+  if (inherits(family, "extended.family")) {
     stop("Unknown family object. Did you use mgcv::gfam()?")
   }
   family_list <- if (!inherits(family, "galamm_extended_family")) {
