### 2025-09-28 16b721b
added residual type option to plot()
