### 2025-09-28 LCBC-UiO/galamm#206 @osorensen
Closed by #228
