### 2025-09-28 LCBC-UiO/galamm#212 @osorensen
Closing this one, as it seems to be hard to do at the moment. Will keep it in mind next time I meet clever student looking for a thesis project :-)
