### 2025-09-28 LCBC-UiO/galamm#217 @osorensen
Closing now, because this particular issue is addressed, but there are larger related issues which I will address in the month to come.
