### 2025-10-01 LCBC-UiO/galamm#216 @osorensen
Indeed, Pearson residuals are probably only useful for Gaussian responses, and not other families. I'll spend some time reading through the literature and see if I can come up with something better.
