### 2025-10-01 ded5f78
Fixed bug related to #214 #215 #216 #230
