+++ inst/REFERENCES.bib
@@ -128,6 +128,19 @@
   address = {San Antonio, TX}
 }
 
+@book{dunnGeneralizedLinearModels2018,
+  title = {Generalized {{Linear Models With Examples}} in {{R}}},
+  author = {Dunn, Peter K. and Smyth, Gordon K.},
+  year = {2018},
+  series = {Springer {{Texts}} in {{Statistics}}},
+  publisher = {Springer},
+  address = {New York, NY},
+  doi = {10.1007/978-1-4419-0118-7},
+  copyright = {http://www.springer.com/tdm},
+  isbn = {978-1-4419-0117-0 978-1-4419-0118-7},
+  keywords = {generalized linear models,likelihood score tests,linear regression,Randomized quantile residuals,Saddlepoint approximation,Tweedie family distribution}
+}
+
 @article{fletcherNewApproachVariable1970,
   title = {A New Approach to Variable Metric Algorithms},
   author = {Fletcher, R.},
