### 2025-10-01 2dca7fc
started adding functinoality to plot.galamm
