+++ tests/testthat/test-galamm-semiparametric.R
@@ -3,18 +3,15 @@
 #' @noRd
 NULL
 
+testthat::skip_if_not_installed("gamm4")
+
 test_that("galamm reproduces gamm4", {
   dat <- subset(cognition, domain == 1 & item == "11" & id < 50)
 
   mod <- galamm(formula = y ~ s(x), data = dat)
-  # Not running gamm4 in tests, but taking outputs from it
-  # mod_comp <- gamm4::gamm4(formula = y ~ s(x), data = dat, REML = FALSE)
+  mod_comp <- gamm4::gamm4(formula = y[, 1] ~ s(x), data = dat, REML = FALSE)
 
-  expect_equal(unname(mod$gam$edf), c(
-    1, 0.917922212177643, 0.835640672846923, 0.473700240730942,
-    -0.0205750525522368, 0.170693269673505, -0.340640339069972, 0.0486758453821619,
-    1.58020490382619, 1.00000000000004
-  ), tolerance = .001)
+  expect_equal(unname(mod$gam$edf), unname(mod_comp$gam$edf), tolerance = .001)
 
   expect_snapshot(print(summary(mod$gam), digits = 2))
 
@@ -22,576 +19,120 @@ test_that("galamm reproduces gamm4", {
   on.exit(dev.off(), add = TRUE)
   expect_type(plot_smooth(mod), "list")
 
-  expect_equal(head(predict(mod)), structure(c(
-    `1` = 0.432267467571946, `4` = 0.850917209622746,
-    `7` = 2.22066429268562, `10` = 2.22494341092419, `13` = 1.74697112222858,
-    `16` = 1.05407575060804
-  ), dim = 6L, dimnames = list(c(
-    "1", "4",
-    "7", "10", "13", "16"
-  ))), tolerance = .01)
+  expect_equal(head(predict(mod)), head(predict(mod_comp$gam)), tolerance = .01)
 
   mod <- galamm(formula = y ~ t2(x), data = dat)
-  # mod_comp <- gamm4::gamm4(formula = y ~ t2(x), data = dat, REML = FALSE)
-  expect_equal(unname(mod$gam$edf), c(
-    1, 0.778128590494658, 0.921997998438558, 0.987589732872169,
-    0.999999999999999
-  ), tolerance = .001)
+  mod_comp <- gamm4::gamm4(formula = y[, 1] ~ t2(x), data = dat, REML = FALSE)
+  expect_equal(unname(mod$gam$edf), unname(mod_comp$gam$edf), tolerance = .001)
 
   expect_snapshot(print(summary(mod$gam), digits = 2))
-  expect_equal(head(predict(mod)), structure(c(
-    `1` = 0.430919582024448, `4` = 0.845539207864478,
-    `7` = 2.25395029739734, `10` = 2.25707170189091, `13` = 1.75570899492922,
-    `16` = 1.05353978411553
-  ), dim = 6L, dimnames = list(c(
-    "1", "4",
-    "7", "10", "13", "16"
-  ))), tolerance = .01)
+  expect_equal(head(predict(mod)), head(predict(mod_comp$gam)), tolerance = .01)
 
   mod <- galamm(formula = y ~ s(x, fx = TRUE) + (1 | id), data = dat)
-  # mod_comp <- gamm4::gamm4(
-  #   formula = y ~ s(x, fx = TRUE),
-  #   random = ~ (1 | id), data = dat, REML = FALSE
-  # )
+  mod_comp <- gamm4::gamm4(
+    formula = y[, 1] ~ s(x, fx = TRUE),
+    random = ~ (1 | id), data = dat, REML = FALSE
+  )
 
-  expect_equal(deviance(mod), 712.425690398506, tolerance = .0001)
+  expect_equal(deviance(mod), deviance(mod_comp$mer), tolerance = .0001)
[…]
