### 2025-11-25 c2ab751
rebuilt vignettes
