+++ vignettes/semiparametric.Rmd
@@ -261,12 +261,12 @@ summary(mod_gamm4$mer)
 #> 
 #> Scaled residuals: 
 #>     Min      1Q  Median      3Q     Max 
-#> -8.2547  0.0786  0.1946  0.4248  0.8213 
+#> -8.2548  0.0786  0.1946  0.4248  0.8213 
 #> 
 #> Random effects:
 #>  Groups Name        Variance Std.Dev.
 #>  id     (Intercept) 0.2306   0.4802  
-#>  Xr     s(x)        0.9694   0.9846  
+#>  Xr     s(x)        0.9695   0.9846  
 #> Number of obs: 1600, groups:  id, 200; Xr, 8
 #> 
 #> Fixed effects:
@@ -608,19 +608,21 @@ mod_byvar <- galamm(
 #> At X0, 0 variables are exactly at the bounds
 #> At iterate     0  f=        25324  |proj g|=       9642.6
 #> At iterate    50  f =        136.9  |proj g|=        716.66
-#> At iterate   100  f =      -1055.1  |proj g|=        2101.3
-#> At iterate   150  f =        -1069  |proj g|=        12.671
+#> At iterate   100  f =      -1036.1  |proj g|=        2364.3
+#> At iterate   150  f =        -1069  |proj g|=          46.6
+#> At iterate   200  f =      -1147.4  |proj g|=         95.18
+#> At iterate   250  f =      -1217.6  |proj g|=        53.837
 #> 
-#> iterations 192
-#> function evaluations 210
-#> segments explored during Cauchy searches 195
+#> iterations 266
+#> function evaluations 292
+#> segments explored during Cauchy searches 269
 #> BFGS updates skipped 0
 #> active bounds at final generalized Cauchy point 0
-#> norm of the final projected gradient 5.66964
-#> final function value -1147.33
+#> norm of the final projected gradient 8.60877
+#> final function value -1217.8
 #> 
-#> F = -1147.33
-#> final  value -1147.325580 
+#> F = -1217.8
+#> final  value -1217.796853 
 #> converged
 ```
 
@@ -636,44 +638,44 @@ summary(mod_byvar)
 #> Control: galamm_control(optim_control = list(factr = 1e+09, trace = 3,      REPORT = 50, maxit = 1000))
 #> 
 #>      AIC      BIC   logLik deviance df.resid 
-#>  -2258.7  -2126.8   1147.3  -2294.7    11182 
+#>  -2399.6  -2267.8   1217.8  -2435.6    11182 
 #> 
 #> Scaled residuals: 
 #>      Min       1Q   Median       3Q      Max 
-#> -2097.34 -1025.85  -872.87    18.58    46.76 
+#> -2115.29 -1034.70  -885.48    21.85    50.66 
 #> 
 #> Lambda:
 #>         ability1       SE ability3        SE
 #> lambda1   1.0000        .        .         .
-#> lambda2   1.3990 0.002570        .         .
-#> lambda3   0.3009 0.002004        .         .
+#> lambda2   1.3986 0.002569        .         .
+#> lambda3   0.3013 0.002007        .         .
 #> lambda4        .        .   1.0000         .
-#> lambda5        .        .   0.9993 0.0007675
-#> lambda6        .        .   1.0001 0.0007677
-#> lambda7        .        .   1.9994 0.0015666
+#> lambda5        .        .   0.9993 0.0007676
+#> lambda6        .        .   1.0001 0.0007679
+#> lambda7        .        .   1.9995 0.0015686
 #> 
 #> Random effects:
 #>  Groups       Name                  Variance Std.Dev. Corr 
[…]
