### 2025-11-25 LCBC-UiO/galamm#209 @osorensen
I think this is out of scope for now, but will keep it in mind as a possible future extension.
