+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: galamm
 Title: Generalized Additive Latent and Mixed Models
-Version: 0.3.0
+Version: 0.3.0.9000
 Authors@R: c(
     person(given = "Øystein",
            family = "Sørensen",
