### 2025-11-26 5781ced
added model frame S3 function
