+++ man/model.frame.galamm.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/model_frame.R
+\name{model.frame.galamm}
+\alias{model.frame.galamm}
+\title{Extract the model frame from a galamm object}
+\usage{
+\method{model.frame}{galamm}(object, ...)
+}
+\arguments{
+\item{object}{An object of type \code{galamm} as described in \link{galammObject}.}
+
+\item{...}{Ignored – present for S3 compatibility.}
+}
+\value{
+A data.frame containing the variables that were used in the fit.
+}
+\description{
+Returns the data frame that was used to fit the model. The function first
+looks for a component named \code{data} (added by the \code{galamm()} function).  If
+that component is missing it falls back to re‑evaluating the expression
+supplied in the original call (\code{call$data}).  This mirrors the behavior
+of \code{model.frame.lmerMod} and other modelling packages.
+}
