### 2025-11-26 c13729a
gave model.frame.galamm a family
