+++ R/model_frame.R
@@ -10,6 +10,7 @@
 #' @param object An object of type `galamm` as described in [galammObject].
 #' @param ...   Ignored – present for S3 compatibility.
 #' @return A data.frame containing the variables that were used in the fit.
+#' @family details of model fit
 #' @export
 model.frame.galamm <- function(object, ...) {
   if (!is.null(object$data)) {
