+++ man/fitted.galamm.Rd
@@ -42,6 +42,7 @@ Other details of model fit:
 \code{\link{formula.galamm}()},
 \code{\link{llikAIC}()},
 \code{\link{logLik.galamm}()},
+\code{\link{model.frame.galamm}()},
 \code{\link{nobs.galamm}()},
 \code{\link{predict.galamm}()},
 \code{\link{print.VarCorr.galamm}()},
