+++ man/logLik.galamm.Rd
@@ -44,6 +44,7 @@ Other details of model fit:
 \code{\link{fixef}()},
 \code{\link{formula.galamm}()},
 \code{\link{llikAIC}()},
+\code{\link{model.frame.galamm}()},
 \code{\link{nobs.galamm}()},
 \code{\link{predict.galamm}()},
 \code{\link{print.VarCorr.galamm}()},
