### 2025-11-26 c732602
fixes for r cmd check
