+++ DESCRIPTION
@@ -55,6 +55,7 @@ Suggests:
     gamm4,
     knitr,
     PLmixed,
+    rlang,
     rmarkdown,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
