+++ reference/print.VarCorr.galamm.md
@@ -71,6 +71,7 @@ Other details of model fit:
 [`formula.galamm()`](https://lcbc-uio.github.io/galamm/reference/formula.galamm.md),
 [`llikAIC()`](https://lcbc-uio.github.io/galamm/reference/llikAIC.md),
 [`logLik.galamm()`](https://lcbc-uio.github.io/galamm/reference/logLik.galamm.md),
+[`model.frame.galamm()`](https://lcbc-uio.github.io/galamm/reference/model.frame.galamm.md),
 [`nobs.galamm()`](https://lcbc-uio.github.io/galamm/reference/nobs.galamm.md),
 [`predict.galamm()`](https://lcbc-uio.github.io/galamm/reference/predict.galamm.md),
 [`ranef.galamm()`](https://lcbc-uio.github.io/galamm/reference/ranef.galamm.md),
