### 2025-11-26 7326c2e
implemented an abline for the plot without conditioning. need to figure out xyplot tomorrow
