### 2025-11-27 9206c17
fixed the plot function
