+++ reference/plot.galamm.md
@@ -10,7 +10,7 @@ function for definition of the residuals being used.
 
 ``` r
 # S3 method for class 'galamm'
-plot(x, residuals = c("pearson", "deviance"), ...)
+plot(x, form = resid(., type = "pearson") ~ fitted(.), abline = NULL, ...)
 ```
 
 ## Arguments
@@ -20,11 +20,15 @@ plot(x, residuals = c("pearson", "deviance"), ...)
   An object of class `galamm` returned from
   [`galamm`](https://lcbc-uio.github.io/galamm/reference/galamm.md).
 
-- residuals:
+- form:
 
-  Character of length one describing the type of residuals to be
-  returned. One of `"pearson"` and `"deviance"`. Argument is case
-  sensitive.
+  An option formula specifying the desired type of plot. Conditioning
+  variables are specified with a vertical bar.
+
+- abline:
+
+  An optional numeric vector specifying the intercept and slope of a
+  line to be added to the plot.
 
 - ...:
 
@@ -33,6 +37,18 @@ plot(x, residuals = c("pearson", "deviance"), ...)
 ## Value
 
 A plot is displayed.
+
+## Details
+
+The interface of this function is designed to be similar to the
+`plot.merMod` function from `lme4` (Bates et al. 2015) .
+
+## References
+
+Bates DM, Mächler M, Bolker B, Walker S (2015). “Fitting Linear
+Mixed-Effects Models Using Lme4.” *Journal of Statistical Software*,
+**67**(1), 1–48. ISSN 1548-7660,
+[doi:10.18637/jss.v067.i01](https://doi.org/10.18637/jss.v067.i01) .
 
 ## See also
 
@@ -48,15 +64,74 @@ Other summary functions:
 [`print.summary.galamm()`](https://lcbc-uio.github.io/galamm/reference/print.summary.galamm.md),
 [`summary.galamm()`](https://lcbc-uio.github.io/galamm/reference/summary.galamm.md)
 
+## Author
+
+Douglas Bates, Martin Maechler, Ben Bolker, and Steven Walker, with
+modifications by Øystein Sørensen.
+
 ## Examples
 
 ``` r
-# Linear mixed model example from lme4
+## Linear mixed model example from lme4
 data("sleepstudy", package = "lme4")
 mod <- galamm(Reaction ~ Days + (Days | Subject), data = sleepstudy)
 
-# Diagnostic plot
+# Diagnostic plot of Pearson residuals versus fitted values
 plot(mod)
 
-plot(mod, residuals = "deviance")
+# Include a straight line at zero
+plot(mod, abline = c(0, 0))
+
+
+# Diagnostic plot of Pearson residuals versus fitted values per subject
+plot(mod, form = resid(., type = "pearson") ~ fitted(.) | Subject)
+
+
+# Residuals plotted against time with a straight line at zero
+plot(mod, form = resid(., type = "pearson") ~ Days, abline = c(0, 0))
+
+
+# Residuals plotted against time per subject with a straight line at zero
+plot(mod, form = resid(., type = "pearson") ~ Days | Subject,
+     abline = c(0, 0))
+
+
+## Logistic mixed model example from lme4
+data("cbpp", package = "lme4")
+mod <- galamm(cbind(incidence, size - incidence) ~ period + (1 | herd),
+              data = cbpp, family = binomial)
+
+# Diagnostic plot using Pearson residuals
+plot(mod)
+
+
+# Diagnostic plot using deviance residuals
+plot(mod, resid(., type = "deviance") ~ fitted(.))
+
+
+# Diagnostic plot per herd
[…]
