### 2025-11-28 adcfe30
added boxplot option
