### 2025-11-28 f62412c
added test and updated documentation
