### 2025-11-28 4f80fbc
added qqmath function
