+++ news.md
@@ -1,5 +1,6 @@
 # galamm 0.3.0 (development versions)
 
+- A qqmath() function has now been added, which makes quantile-quantile plots.
 - residuals.galamm() now has a "scaled = TRUE/FALSE" argument.
 - plot.galamm() has been considerably extended.
 
