### 2025-11-28 LCBC-UiO/galamm#216 @osorensen
An update on this one: I have now extended the `residuals.galamm()` function so that it can give deviance residuals and the user can choose whether to scale them or not. In addition, the diagnostic plot function `plot.galamm()` now has functionality quite similar to `lme4::plot.merMod`, so that it can give a range of different diagnostics which include these residuals. I hence hope this issue is addressed as far as the rOpenSci revision is concerned.
