### 2025-11-28 3710abe
added diagnostics to latent observed interaction model
