### 2025-11-28 f2d7cf3
updated diagnostics in semiparametric model
