### 2025-11-28 946d8f7
reran vignettes
