### 2025-12-03 LCBC-UiO/galamm#220 @osorensen
I'll add documentation covering how do do posterior sampling. After some testing, I notice that mgcv::gam.mh does not seem to work well gamm4::gamm4 or mgcv::gamm, and hence also cannot be used out-of-the-box for galamm::galamm. I hence do it the easy way and support direct sampling from the empirical Bayes posterior, which is what was done in the Psychometrika paper, rather than the potentially more accurate random walk Metropolis sampler implemented in mgcv::gam.mh.
