### 2025-12-10 90e817b
solves #248
