### 2025-12-10 836ccd3
fixed a formula bug
