+++ reference/predict.galamm.md
@@ -10,7 +10,7 @@ predictors is supported.
 
 ``` r
 # S3 method for class 'galamm'
-predict(object, newdata = NULL, type = c("link", "response"), ...)
+predict(object, newdata = NULL, type = c("link", "response", "lpmatrix"), ...)
 ```
 
 ## Arguments
@@ -96,4 +96,18 @@ predict(count_mod, newdata = nd, type = "response")
 #>       [,1]
 #> 1 8.917850
 #> 2 6.248365
+
+# Semiparametric model
+dat <- subset(cognition, domain == 1 & item == "11")
+dat$y <- dat$y[, 1]
+mod <- galamm(y ~ s(x) + (1 | id), data = dat)
+
+# Get the linear predictor matrix
+Xp <- predict(mod, type = "lpmatrix")
+# Compute the estimated smooth function
+# See the vignette on posterior sampling for details
+betas <- coef(mod$gam)
+fit <- Xp %*% betas
+plot(dat$x, fit)
+
 ```
