### 2025-12-10 15a992e
made vignette on posterior sampling
