### 2025-12-10 LCBC-UiO/galamm#220 @osorensen
I have now made a vignette which shows how one can do posterior sampling in galamm. The vignette is pretty short, because the main goal with it is to show which functions in galamm can be used to do posterior sampling. It references to a somewhat longer tutorial in Gavin Simpson's blog as well as the relevant section of Simon Wood's 2017 book.
