+++ llms.txt
@@ -450,6 +450,8 @@ Functions to aid in optimization.
 
 ### Advanced topics
 
+- [Posterior
+  Sampling](https://lcbc-uio.github.io/galamm/articles/posterior_sampling.md):
 - [Computational
   Scaling](https://lcbc-uio.github.io/galamm/articles/scaling.md):
 - [Optimization](https://lcbc-uio.github.io/galamm/articles/optimization.md):
