### 2025-12-11 e43e8ee
implemented basic draw function
