+++ man/draw.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gratia_wrappers.R
+\name{draw}
+\alias{draw}
+\alias{draw.default}
+\alias{draw.galamm}
+\title{Draw method for galamm objects}
+\usage{
+draw(object, ...)
+
+\method{draw}{default}(object, ...)
+
+\method{draw}{galamm}(object, ...)
+}
+\value{
+A ggplot object.
+}
+\description{
+This function uses \code{gratia::draw} to visualize smooth terms. See
+\code{\link[gratia:draw]{gratia::draw()}} for details. When \code{object} is not of class
+\code{galamm}, it is forwarded to \code{gratia::draw()}.
+}
+\examples{
+dat <- subset(cognition, domain == 1 & item == "11")
+dat$y <- dat$y[, 1]
+mod <- galamm(y ~ s(x) + (1 | id), data = dat)
+
+draw(mod)
+
+}
+\seealso{
+Other summary functions: 
+\code{\link{anova.galamm}()},
+\code{\link{plot_smooth.galamm}()},
+\code{\link{print.galamm}()},
+\code{\link{print.summary.galamm}()},
+\code{\link{summary.galamm}()}
+}
+\concept{summary functions}
