+++ NAMESPACE
@@ -5,6 +5,8 @@ S3method(anova,galamm)
 S3method(coef,galamm)
 S3method(confint,galamm)
 S3method(deviance,galamm)
+S3method(draw,default)
+S3method(draw,galamm)
 S3method(extract_optim_parameters,galamm)
 S3method(factor_loadings,galamm)
 S3method(family,galamm)
@@ -27,6 +29,7 @@ S3method(sigma,galamm)
 S3method(summary,galamm)
 S3method(vcov,galamm)
 export(VarCorr)
+export(draw)
 export(extract_optim_parameters)
 export(factor_loadings)
 export(fixef)
