+++ DESCRIPTION
@@ -33,6 +33,7 @@ URL: https://github.com/LCBC-UiO/galamm, https://lcbc-uio.github.io/galamm/
 BugReports: https://github.com/LCBC-UiO/galamm/issues
 Encoding: UTF-8
 Imports: 
+    gratia,
     lattice,
     lme4,
     Matrix,
