### 2025-12-11 b8ef7a1
found a better solution
