+++ man/draw.Rd
@@ -1,15 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gratia_wrappers.R
-\name{draw}
-\alias{draw}
-\alias{draw.default}
+% Please edit documentation in R/gratia_draw.R
+\name{draw.galamm}
 \alias{draw.galamm}
 \title{Draw method for galamm objects}
 \usage{
-draw(object, ...)
-
-\method{draw}{default}(object, ...)
-
 \method{draw}{galamm}(object, ...)
 }
 \value{
