+++ man/summary.galamm.Rd
@@ -60,7 +60,7 @@ the generic.
 
 Other summary functions: 
 \code{\link{anova.galamm}()},
-\code{\link{draw}()},
+\code{\link{draw.galamm}()},
 \code{\link{plot_smooth.galamm}()},
 \code{\link{print.galamm}()},
 \code{\link{print.summary.galamm}()}
