+++ NAMESPACE
@@ -5,7 +5,6 @@ S3method(anova,galamm)
 S3method(coef,galamm)
 S3method(confint,galamm)
 S3method(deviance,galamm)
-S3method(draw,default)
 S3method(draw,galamm)
 S3method(extract_optim_parameters,galamm)
 S3method(factor_loadings,galamm)
@@ -46,6 +45,7 @@ export(t2)
 export(t2l)
 importFrom(Rcpp,sourceCpp)
 importFrom(Rdpack,reprompt)
+importFrom(gratia,draw)
 importFrom(lattice,qqmath)
 importFrom(mgcv,s)
 importFrom(mgcv,t2)
