### 2025-12-11 802427b
made derivatives function
