+++ man/VarCorr.Rd
@@ -43,6 +43,7 @@ as.data.frame(VarCorr(mod))
 Other details of model fit: 
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
+\code{\link{derivatives.galamm}()},
 \code{\link{deviance.galamm}()},
 \code{\link{factor_loadings.galamm}()},
 \code{\link{family.galamm}()},
