+++ man/deviance.galamm.Rd
@@ -38,6 +38,7 @@ Other details of model fit:
 \code{\link{VarCorr}()},
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
+\code{\link{derivatives.galamm}()},
 \code{\link{factor_loadings.galamm}()},
 \code{\link{family.galamm}()},
 \code{\link{fitted.galamm}()},
