### 2025-12-11 0d7008c
created appraise method
