+++ man/llikAIC.Rd
@@ -20,6 +20,7 @@ This function is assembles the values used by \code{\link{summary.galamm}}.
 \seealso{
 Other details of model fit: 
 \code{\link{VarCorr}()},
+\code{\link{appraise.galamm}()},
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
 \code{\link{derivatives.galamm}()},
