+++ man/confint.galamm.Rd
@@ -48,6 +48,7 @@ more generally, and \code{\link[=vcov.galamm]{vcov.galamm()}} for the variance-c
 
 Other details of model fit: 
 \code{\link{VarCorr}()},
+\code{\link{appraise.galamm}()},
 \code{\link{coef.galamm}()},
 \code{\link{derivatives.galamm}()},
 \code{\link{deviance.galamm}()},
