+++ man/residuals.galamm.Rd
@@ -56,6 +56,7 @@ function is \code{\link[=residuals]{residuals()}}.
 
 Other details of model fit: 
 \code{\link{VarCorr}()},
+\code{\link{appraise.galamm}()},
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
 \code{\link{derivatives.galamm}()},
