+++ man/model.frame.galamm.Rd
@@ -20,6 +20,7 @@ Extract the model frame from a galamm object
 \seealso{
 Other details of model fit: 
 \code{\link{VarCorr}()},
+\code{\link{appraise.galamm}()},
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
 \code{\link{derivatives.galamm}()},
