+++ man/nobs.galamm.Rd
@@ -31,6 +31,7 @@ nrow(sleepstudy)
 \seealso{
 Other details of model fit: 
 \code{\link{VarCorr}()},
+\code{\link{appraise.galamm}()},
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
 \code{\link{derivatives.galamm}()},
