### 2025-12-11 a8edf9b
made tests of gratia functions
