+++ R/gratia_derivatives.R
@@ -23,7 +23,7 @@ NULL
 #' draw(dd)
 #'
 derivatives.galamm <- function(object, ...) {
-  if (is.null(object$gam)) {
+  if (is.null(object$gam) || length(object$gam) == 0) {
     stop("No GAM object")
   }
   gratia::derivatives(object$gam, ...)
