### 2025-12-11 05e2034
made news and incremented dev version
