+++ man/appraise.galamm.Rd
@@ -1,35 +1,39 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gratia_derivatives.R
-\name{derivatives.galamm}
-\alias{derivatives.galamm}
-\title{Derivatives of estimated smooth via finite differences}
+% Please edit documentation in R/gratia_appraise.R
+\name{appraise.galamm}
+\alias{appraise.galamm}
+\alias{appraise}
+\title{Gratia style model diagnostic plots}
 \usage{
-\method{derivatives}{galamm}(object, ...)
+\method{appraise}{galamm}(model, ...)
+}
+\arguments{
+\item{model}{An object of class \code{galamm} returned from
+\code{\link{galamm}}.}
+
+\item{...}{Other arguments passed on to \code{\link[gratia:appraise]{gratia::appraise()}}.}
 }
 \value{
-A tibble.
+A ggplot object.
 }
 \description{
-This function uses \code{gratia::derivatives} to compute derivatives of
-estimated smooth via finite differences. See
-\code{\link[gratia:derivatives]{gratia::derivatives()}} for details. When \code{object} is not of class
-\code{galamm}, it is forwarded to \code{gratia::derivatives()}.
+This function uses \code{gratia::appraise} to make model diagnostic plots.
+See \code{\link[gratia:appraise]{gratia::appraise()}} for details. When \code{model} is not of class
+\code{galamm}, it is forwarded to \code{gratia::appraise()}.
 }
 \examples{
 dat <- subset(cognition, domain == 1 & item == "11")
 dat$y <- dat$y[, 1]
 mod <- galamm(y ~ s(x) + (1 | id), data = dat)
-
-dd <- derivatives(mod)
-draw(dd)
+appraise(mod)
 
 }
 \seealso{
 Other details of model fit: 
 \code{\link{VarCorr}()},
-\code{\link{appraise.galamm}()},
 \code{\link{coef.galamm}()},
 \code{\link{confint.galamm}()},
+\code{\link{derivatives.galamm}()},
 \code{\link{deviance.galamm}()},
 \code{\link{factor_loadings.galamm}()},
 \code{\link{family.galamm}()},
