### 2025-12-12 LCBC-UiO/galamm#208 @osorensen
I like the idea, but won't have capacity to do this until 2026. I'll tag it is an "enhancement" for now.
