### 2025-12-12 #615 @osorensen
👋 @ldecicco-USGS @nicholasjclark @dill, thanks again for your comments, all of which were very good and important. I have now made my best attempt at addressing them. An overview of the issues can be found [here](https://github.com/LCBC-UiO/galamm/issues/203). The version currently in the main branch of the [GitHub repo](https://github.com/LCBC-UiO/galamm) contains all the changes. Looking forward to hearing back from you.
