+++ reference/logLik.galamm.md
@@ -32,8 +32,10 @@ function.
 
 Other details of model fit:
 [`VarCorr()`](https://lcbc-uio.github.io/galamm/reference/VarCorr.md),
+[`appraise.galamm()`](https://lcbc-uio.github.io/galamm/reference/appraise.galamm.md),
 [`coef.galamm()`](https://lcbc-uio.github.io/galamm/reference/coef.galamm.md),
 [`confint.galamm()`](https://lcbc-uio.github.io/galamm/reference/confint.galamm.md),
+[`derivatives.galamm()`](https://lcbc-uio.github.io/galamm/reference/derivatives.galamm.md),
 [`deviance.galamm()`](https://lcbc-uio.github.io/galamm/reference/deviance.galamm.md),
 [`factor_loadings.galamm()`](https://lcbc-uio.github.io/galamm/reference/factor_loadings.galamm.md),
 [`family.galamm()`](https://lcbc-uio.github.io/galamm/reference/family.galamm.md),
