+++ reference/plot_smooth.galamm.md
@@ -37,6 +37,7 @@ edition. Chapman and Hall/CRC.
 
 Other summary functions:
 [`anova.galamm()`](https://lcbc-uio.github.io/galamm/reference/anova.galamm.md),
+[`draw.galamm()`](https://lcbc-uio.github.io/galamm/reference/draw.galamm.md),
 [`print.galamm()`](https://lcbc-uio.github.io/galamm/reference/print.galamm.md),
 [`print.summary.galamm()`](https://lcbc-uio.github.io/galamm/reference/print.summary.galamm.md),
 [`summary.galamm()`](https://lcbc-uio.github.io/galamm/reference/summary.galamm.md)

[…]
