+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   posterior_sampling: posterior_sampling.html
   scaling: scaling.html
   semiparametric: semiparametric.html
-last_built: 2025-12-21T12:45Z
+last_built: 2025-12-21T12:48Z
 urls:
   reference: https://lcbc-uio.github.io/galamm/reference
   article: https://lcbc-uio.github.io/galamm/articles
