### 2025-12-21 9a9332c
recompiled vignettes
