+++ vignettes/lmm_heteroscedastic.Rmd
@@ -93,6 +93,11 @@ We can confirm that the lme function from the nlme package gives the same result
 
 ``` r
 library(nlme)
+#> 
+#> Attaching package: 'nlme'
+#> The following object is masked from 'package:lme4':
+#> 
+#>     lmList
 mod_nlme <- lme(y ~ x,
   data = hsced, random = list(id = ~1),
   weights = varIdent(form = ~ 1 | item), method = "ML"
