+++ vignettes/posterior_sampling.Rmd
@@ -48,7 +48,7 @@ The estimated smooth function looks like this:
 plot_smooth(mod)
 ```
 
-![plot of chunk unnamed-chunk-66](unnamed-chunk-66-1.png)
+![plot of chunk unnamed-chunk-3](unnamed-chunk-3-1.png)
 
 By making a grid, we can find where it reaches its maximum:
 
@@ -108,7 +108,7 @@ plot(grid$x, fits[, 1], type = "l")
 for(i in sample(2:1000, 10)) lines(grid$x, fits[, i])
 ```
 
-![plot of chunk unnamed-chunk-73](unnamed-chunk-73-1.png)
+![plot of chunk unnamed-chunk-10](unnamed-chunk-10-1.png)
 
 We can now get a posterior 95% interval for the position of the maximum by simply looking at percentiles.
 
@@ -117,7 +117,7 @@ We can now get a posterior 95% interval for the position of the maximum by simpl
 maxima <- grid$x[apply(fits, 2, which.max)]
 quantile(maxima, probs = c(.025, .975))
 #>     2.5%    97.5% 
-#> 0.488975 0.568000
+#> 0.491975 0.565000
 ```
 
 We can also visualize the posterior distribution.
@@ -127,7 +127,7 @@ We can also visualize the posterior distribution.
 hist(maxima)
 ```
 
-![plot of chunk unnamed-chunk-75](unnamed-chunk-75-1.png)
+![plot of chunk unnamed-chunk-12](unnamed-chunk-12-1.png)
 
 
 # References
