+++ articles/glmm_factor.md
@@ -412,6 +412,7 @@ flexibility we need.
 
 ``` r
 library(lme4)
+#> Loading required package: Matrix
 count_mod_lme4 <- glmer(
   formula = y ~ lbas * treat + lage + v4 + (1 | subj),
   data = epilep,
@@ -457,7 +458,7 @@ summary(count_mod_lme4)
 #> (Intercept)   1.7937     0.1045  17.165  < 2e-16 ***
 #> lbas          0.8844     0.1308   6.764 1.34e-11 ***
 #> treat        -0.3349     0.1474  -2.272  0.02311 *  
-#> lage          0.4845     0.3458   1.401  0.16122    
+#> lage          0.4845     0.3458   1.401  0.16121    
 #> v4           -0.1611     0.0543  -2.966  0.00301 ** 
 #> lbas:treat    0.3382     0.2025   1.671  0.09479 .  
 #> ---
