### 2026-02-10 #615 @emilyriederer
Hi @nicholasjclark and @dill ! I have taken over as rotation EiC for rOpenSci this quarter and am evaluating our currently open issues. I notice that this is waiting final reviewer approval. At your convenience, could you please review @osorensen 's changes and see if you feel that your comments are sufficiently addessed? You may give final signoff using our [approval template](https://devguide.ropensci.org/approval2template.html).
