### 2026-05-03 #615 @maurolepore
Dear @osorensen, this is to mark the start of my new EiC rotation. Thank you for the thorough revisions back in December, and thanks @dill for the approval.

A brief summary of what I see:
- You addressed both reviews in December.
- @dill approved on 2026-02-17.
- @nicholasjclark hasn't responded yet.

I know @noamross has been trying to reach out via email recently, so the submission is in good hands and moving forwards as indicated in our dev guide:

> If the reviewer has already submitted their primary review, and another reviewer is active and providing substantial feedback, the editor can proceed with the review process, and should take the role of the absent reviewer in determining if the authors' changes are sufficient.
> — https://devguide.ropensci.org/editorguide#during-review

Thanks!
