+++ reference/print.summary.galamm.md
@@ -34,7 +34,7 @@ Summary printed to screen. Invisibly returns the argument `x`.
 
 Bates DM, Mächler M, Bolker B, Walker S (2015). “Fitting Linear
 Mixed-Effects Models Using Lme4.” *Journal of Statistical Software*,
-**67**(1), 1–48. ISSN 1548-7660,
+**67**(1), 1–48. ISSN 1548-7660.
 [doi:10.18637/jss.v067.i01](https://doi.org/10.18637/jss.v067.i01) .
 
 ## See also
