### 2026-06-03 5b7fe23
Update VarCorr snapshots for lme4 output
