+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   posterior_sampling: posterior_sampling.html
   scaling: scaling.html
   semiparametric: semiparametric.html
-last_built: 2026-06-03T07:59Z
+last_built: 2026-06-03T08:41Z
 urls:
   reference: https://lcbc-uio.github.io/galamm/reference
   article: https://lcbc-uio.github.io/galamm/articles
